% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abort_missing_ref}
\alias{abort_missing_ref}
\title{Abort if relevant environment variables are not set}
\usage{
abort_missing_ref()
}
\description{
This function is only exported because it is called when environment
variables are not used.
@export
}
