% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{activate}
\alias{activate}
\alias{deactivate}
\title{Activate touchstone environment}
\usage{
activate(
  head_branch = gert::git_branch(),
  base_branch = getOption("touchstone.default_base_branch", "main"),
  n = 1,
  env = parent.frame()
)

deactivate(env = parent.frame())
}
\arguments{
\item{head_branch}{Git branch to be used as the \code{GITHUB_HEAD_REF} branch
(i.e. the branch with new changes) when running benchmarks. Defaults to the
current branch.}

\item{base_branch}{Git branch for the \code{GITHUB_BASE_REF} (i.e. the branch you
want to merge your changes into) when running benchmarks. Defaults to 'main'
if the option \code{touchstone.default_base_branch}is not set.}

\item{n}{Number of times benchmarks should be run for each \code{branch}. Will
override \code{n} argument in all interactive calls to \link{benchmark_run}}

\item{env}{In which environment the temporary changes should be made.
For use within functions.}
}
\description{
This sets environment variables, R options and library paths to work
interactively on the \link{touchstone_script}.
}
\section{Functions}{
\itemize{
\item \code{deactivate}: Restore the original environment state.
}}

\examples{
\dontrun{
activate()
# You can now test parts of your touchstone script, e.g. touchstone/script.R
deactivate()
}
}
