% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScans.R
\name{addMasterScanInfo}
\alias{addMasterScanInfo}
\title{Add information from the master scan file to an MS3 search result}
\usage{
addMasterScanInfo(ms3results, masterScanFile)
}
\arguments{
\item{ms3results}{ms3 level results table as parsed by \code{readMS3results()}}

\item{masterScanFile}{master scan file as created by \code{createMasterScanFile()}}
}
\value{
A data frame
}
\description{
Adds information from scan-dependencies generated with \code{createMasterScanFile()}.
}
\seealso{
\code{\link[=createMasterScanFile]{createMasterScanFile()}},\code{\link[=readMS3results]{readMS3results()}}
}
