% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{addXlinkCount}
\alias{addXlinkCount}
\title{count number of CSMs per URP and number of URPs per PP}
\usage{
addXlinkCount(datTab)
}
\arguments{
\item{datTab}{}
}
\description{
count number of CSMs per URP and number of URPs per PP
}
