% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{assignModules}
\alias{assignModules}
\title{Assigns crosslinks to modules defined in a Touchstone formatted module file}
\usage{
assignModules(datTab, moduleFile)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{moduleFile}{Parsed, Touchstone format module file}
}
\value{
a data frame
}
\description{
Assigns crosslinks to modules defined in a Touchstone formatted module file
}
