% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{assignXLinkClass}
\alias{assignXLinkClass}
\title{Determines whether crosslink is inter-protein (hetromeric) or intra-protein (homomeric)}
\usage{
assignXLinkClass(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
Determines whether crosslink is inter-protein (hetromeric) or intra-protein (homomeric)
}
