% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{benchmark_run_iteration}
\alias{benchmark_run_iteration}
\title{Run a benchmark iteration}
\usage{
benchmark_run_iteration(
  expr_before_benchmark,
  dots,
  ref,
  block,
  n = getOption("touchstone.n_iterations", 1)
)
}
\arguments{
\item{expr_before_benchmark}{Expression to run before
the benchmark is ran, will be captured with \code{\link[rlang:nse-defuse]{rlang::enexpr()}}.
So you can use quasiquotation.}

\item{dots}{list of quoted expressions (length 1).}

\item{ref}{A character vector of length one to indicate the git ref (i.e.
commit, tag, branch etc) of the benchmarking.}

\item{block}{All refs appear once in a block.}

\item{n}{Number of iterations to run a benchmark within an iteration.}
}
\description{
Run a benchmark iteration
}
\keyword{internal}
