% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{benchmark_run_ref}
\alias{benchmark_run_ref}
\title{Run a benchmark for git refs}
\usage{
benchmark_run_ref(
  refs,
  expr_before_benchmark,
  ...,
  n = 20,
  path_pkg = ".",
  install_quick = TRUE,
  install_dependencies = FALSE
)
}
\arguments{
\item{refs}{Git refs passed as \code{ref} to \code{\link[=benchmark_iteration_prepare]{benchmark_iteration_prepare()}}.}

\item{expr_before_benchmark}{Character vector with code to run before
the benchmark is ran, will be evaluated with \code{\link[=exprs_eval]{exprs_eval()}}.}

\item{...}{Named character vector of length one with code to benchmark, will
be evaluated with \code{\link[=exprs_eval]{exprs_eval()}}.}

\item{n}{Number of time to run benchmark.}

\item{path_pkg}{The path to the repository to install.}

\item{install_quick}{Passed to \code{\link[devtools:install]{devtools::install()}} as \code{quick}.}

\item{install_dependencies}{Passed to \code{\link[devtools:install]{devtools::install()}}. Set to \code{FALSE}
can help when ran locally without internet connection.}
}
\description{
Run a benchmark for git refs
}
\details{
Runs the following loop \code{n} times:
\itemize{
\item Installs random branch from \code{refs} of the package \code{path_profiling_pkg}.
\item runs setup code \code{exp_before_ref}.
\item benchmarks \code{expr_to_benchmark}.
Returns all timings.
}
}
