% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{benchmark_verbalize}
\alias{benchmark_verbalize}
\title{Create nice text from benchmarks}
\usage{
benchmark_verbalize(benchmark, timings, refs, ci)
}
\value{
A character vector with the text that verbalizes the benchmark.
}
\description{
\code{refs} must be passed because the order is relevant.
}
\keyword{internal}
