% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{benchmarks_analyse}
\alias{benchmarks_analyse}
\title{Turn raw benchmark results into text and figures}
\usage{
benchmarks_analyse(
  refs = c(Sys.getenv("GITHUB_BASE_REF"), Sys.getenv("GITHUB_HEAD_REF"))
)
}
\description{
Creates two side effects:
}
\details{
\itemize{
\item Density plots for each element in \code{refs} are written to \code{touchstone/plots}.
\item A text explaining the speed diff is written to
\code{touchstone/pr-comment/info.txt} for every registered benchmarking
expression.
}
}
