% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{bestPair}
\alias{bestPair}
\title{Parent function for summarization. CLMS data is grouped by summarization variable
and the best scoring CSM from the group is chosen as the representative}
\usage{
bestPair(
  datTab,
  summarizationVar = xlinkedResPair,
  classifier = SVM.score,
  retainGroups = TRUE
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{summarizationVar}{variable to summarize on}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}

\item{retainGroups}{Should pre-existing grouping of datTab be retained?}
}
\value{
A data frame
}
\description{
Parent function for summarization. CLMS data is grouped by summarization variable
and the best scoring CSM from the group is chosen as the representative
}
\keyword{internal}
