% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{bestProtPair}
\alias{bestProtPair}
\title{Summarize on Protein Pairs (PPs)}
\usage{
bestProtPair(datTab, classifier = SVM.score, retainGroups = TRUE)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}

\item{retainGroups}{Will retain pre-existing group in the data frame when apply the summarization.}
}
\value{
A data frame
}
\description{
The default behaviour is to summarize protein-pairs within existing groups.
This would typically be different experimental conditions. To overwrite and
calculate the protein-pairs on the entire data table seq the \code{retainGroup} option
to \code{False}.
}
\seealso{
\code{\link[=bestResPair]{bestResPair()}}, \code{\link[=bestPepPair]{bestPepPair()}}, \code{\link[=bestModPair]{bestModPair()}}
}
