% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{branch_get_or_fail}
\alias{branch_get_or_fail}
\title{Get the branch from the environment variable or fail if not set}
\usage{
branch_get_or_fail(var)
}
\arguments{
\item{var}{The environment variable to retrieve.}
}
\value{
Returns a character vector of length one with the \code{branch} retrieved from the
environment variable \code{var}.
}
\description{
This function is only exported because it is a default argument.
}
