% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{buildClassifier}
\alias{buildClassifier}
\title{Train an SVM scoring function for CLMS data}
\usage{
buildClassifier(
  datTab,
  params = params.best,
  preFilterER = NA,
  scoreName = "SVM.score",
  scalingFactor = the$decoyScalingFactor,
  sampleNo = 20000,
  ...
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{params}{Character vector specifying names of the features in \code{datTab} used to train model.}

\item{preFilterER}{Desired error rate (FDR) for pre-filtering.}

\item{scoreName}{Name for the new scoring function.}

\item{scalingFactor}{An integer k. he multiple by which decoy DB is larger than target DB}

\item{sampleNo}{Size of the training dataset (integer).}

\item{...}{paramters passed down to \code{e1071::svm()}}
}
\value{
A data frame, one column larger than the input containing the new score.
}
\description{
replaced by buildSVM, but keeping it around for backward capability.
\code{buildClassifier()} is the basic function used to create an SVM scoring model
for CLMS results. Useful for testing different sets of parameters. Results
are pre-filtered to a specified FDR (based on Score.Diff) prior to training.
}
\seealso{
\code{\link[=trainClassifier]{trainClassifier()}}
}
