% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{cache_up_to_date}
\alias{cache_up_to_date}
\alias{cache_update}
\alias{cache_get}
\title{Cache package sources within a session}
\usage{
cache_up_to_date(ref, path_pkg)

cache_update(ref, path_pkg)

cache_get()
}
\arguments{
\item{ref}{A reference to a git commit. Currently, only branch names are
supported.}

\item{path_pkg}{The path to the repository to install.}
}
\description{
This is required to make sure \code{\link[remotes:install_local]{remotes::install_local()}} installs again
when source code changed.
}
\keyword{internal}
