% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculateCrossDiagnosticPair}
\alias{calculateCrossDiagnosticPair}
\title{function for sequential scan data, e.g. sequential MS2-CID and MS2-ETD determined triggered on the same precursor ion.}
\usage{
calculateCrossDiagnosticPair(datTab.etd, datTab.cid, masterScanFile)
}
\arguments{
\item{datTab.etd}{parsed search table that is not expected to contain diagnostic ions (does not necessarily need to be ETD)}

\item{datTab.cid}{parsed search table expected to contain diagnostic ions (typical MS2-CID)}

\item{masterScanFile}{master scan file that charts the dependencies between different scans.}
}
\description{
function for sequential scan data, e.g. sequential MS2-CID and MS2-ETD determined triggered on the same precursor ion.
}
