% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculateCrossDiagnosticPairs}
\alias{calculateCrossDiagnosticPairs}
\title{Determines if diagnostic/reporter ions from cleavable crosslinkers are present
across linked MS2 scans.}
\usage{
calculateCrossDiagnosticPairs(datTab.etd, datTab.cid, masterScanFile)
}
\arguments{
\item{datTab.etd}{Parsed CLMS table that is not expected to contain diagnostic ions (does not need to be ETD)}

\item{datTab.cid}{Parsed CLMS table that is expected to contain diagnostic ions (typical MS2-CID)}

\item{masterScanFile}{Master scan file that charts the dependencies between different scans.}
}
\value{
A data frame
}
\description{
For instance, when sequential MS2-CID and MS2-ETD are triggered on the same precursor ion.
\code{calculateCrossDiagnosticPairs()} looks for the \verb{MH*/#-ion} in the CID scan, but
annotated it as having been found in both scans. Requires a \code{masterScanFile} typically
generated from PAVA processed mgf files.
}
\details{
This function is mostly intended for internal testing and development.
}
\seealso{
\code{\link[=calculateDiagnosticPairs]{calculateDiagnosticPairs()}}
}
