% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{calculateDecoyFractions}
\alias{calculateDecoyFractions}
\title{Estimates fraction of TT hits with one or both peptides incorrectly identified.}
\usage{
calculateDecoyFractions(datTab, scalingFactor = the$decoyScalingFactor)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{scalingFactor}{An integer k. The multiple by which decoy DB is larger than target DB}
}
\value{
A numeric vector
}
\description{
Internal function called by other functions in \code{fdr.R}
}
\keyword{internal}
