% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculateDecoys}
\alias{calculateDecoys}
\title{Assign decoy/target status for each line of a crosslinked dataframe.}
\usage{
calculateDecoys(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
Assign decoy/target status for each line of a crosslinked dataframe.
}
