% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculateDiagnosticPairs}
\alias{calculateDiagnosticPairs}
\title{Determines which crosslinks contain matches to diagnostic/reporter ions (MH*/#)
from cleavable crosslinking reagents.}
\usage{
calculateDiagnosticPairs(datTab, int.thresh = 0.05)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{int.thresh}{Minimum intensity for diagnostic ion.}
}
\value{
A data frame
}
\description{
Determines which crosslinks contain matches to diagnostic/reporter ions (MH*/#)
from cleavable crosslinking reagents.
}
\seealso{
\code{\link[=getDiagnosticMatches]{getDiagnosticMatches()}}
}
