% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{calculateFDR.unseparated}
\alias{calculateFDR.unseparated}
\title{Calculate FDR for a dataset, above a single score threshold}
\usage{
calculateFDR.unseparated(
  datTab,
  threshold = -100,
  classifier = "SVM.score",
  scalingFactor = the$decoyScalingFactor
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{threshold}{Score threshold.}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}

\item{scalingFactor}{An integer k. The multiple by which decoy DB is larger than target DB}
}
\value{
A numeric value
}
\description{
\code{calculateFDR.unseparated()} estimates the FDR for a dataset based on counting the number
of decoy hits. Does not separate inter- and intra-protein hits.
}
\seealso{
\code{\link[=calculateFDR]{calculateFDR()}}
}
\keyword{internal}
