% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{calculateGroundTruthFDR}
\alias{calculateGroundTruthFDR}
\title{Calculate actual FDR when analyzing benchmarking datasets where ground truth is established.}
\usage{
calculateGroundTruthFDR(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
Requires the dataset to have a logical column named \code{groundTruth}. This is typically
for testing against benchmarking datasets. This particular function was written
to reproduce a specific analysis where each \code{Fraction} contained a replicate.
Will turn this into a more generally useful function at some point.
}
