% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{calculateHits}
\alias{calculateHits}
\title{Calculates number of interProtein and intraProtein hits above a single threshold.}
\usage{
calculateHits(datTab, threshold = -100, classifier = "SVM.score")
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{threshold}{Single score threshold}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}
}
\value{
A numeric vector
}
\description{
Internal function called by other functions in \code{fdr.R}
}
\keyword{internal}
