% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculatePeptideLengths}
\alias{calculatePeptideLengths}
\title{Calculate length of each peptide in amino acid residues.}
\usage{
calculatePeptideLengths(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
Calculate length of each peptide in amino acid residues.
}
