% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculatePercentMatched}
\alias{calculatePercentMatched}
\title{If MS-Product reporting of \%TIC matched is present use that.  Otherwise approximate with number of matched ions.}
\usage{
calculatePercentMatched(datTab)
}
\arguments{
\item{datTab}{}
}
\description{
If MS-Product reporting of \%TIC matched is present use that.  Otherwise approximate with number of matched ions.
}
