% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScans.R
\name{calculatePrecMass}
\alias{calculatePrecMass}
\title{Calculate precursor mass from m/z and charge}
\usage{
calculatePrecMass(pepMass, charge)
}
\arguments{
\item{pepMass}{m/z from mgf header.}

\item{charge}{z from mgf header.}
}
\value{
A numeric value
}
\description{
Calculate precursor mass from m/z and charge
}
