% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculateProductIons}
\alias{calculateProductIons}
\title{Calculates the number of distinct backbone bond-cleavages observed as in the product ion spectrum.}
\usage{
calculateProductIons(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
N-terminal ions (b-, c-) and C-terminal ions (y-, z-) are translated into a bond index.
It is the number of unique bond indicies that is calculated here.
}
