% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{clearAboveDiag}
\alias{clearAboveDiag}
\title{Internal function used by some Touchstone plots to only display one half of the plot
in symmetrical square matrix representations of cross-linking data.}
\usage{
clearAboveDiag(sqMatrix)
}
\arguments{
\item{sqMatrix}{A square matrix}
}
\value{
A matrix
}
\description{
Internal function used by some Touchstone plots to only display one half of the plot
in symmetrical square matrix representations of cross-linking data.
}
\keyword{internal}
