% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScans.R
\name{createMasterScanFile}
\alias{createMasterScanFile}
\title{Create a master scan file which maps scan dependencies in mgf peakists.}
\usage{
createMasterScanFile(
  ms3PAVApeaklist,
  ms2PAVApeaklistCID,
  ms2PAVApeaklistETD = NA
)
}
\arguments{
\item{ms3PAVApeaklist}{path to MS3 peaklist.}

\item{ms2PAVApeaklistCID}{path to MS2 CID peaklist (parent of MS3 scans).}

\item{ms2PAVApeaklistETD}{path to MS2 ETD peaklist (optional).}
}
\value{
A data frame
}
\description{
\code{createMasterScanFile()} is for scan sequences that contain an MS3 scan. May
contain an additional MS2 scan in the acquisition sequence (assumed to be ETD
in this function, but can be any). Requires using the PAVA formating conventions
for the peaklists filenames. For instance, PAVA creates separate peaklist
files for each scan type and appends a suffic such as: \verb{_FTMSms2ethcd.txt}
for high-res MS2-EThcD scans, or \verb{_ITMSms3cid.txt} for low-res MS3-CID scans.
The base filename is the same for all the linked peaklist files, corresponding
to the name of the .raw file.
}
\seealso{
\code{\link[=getIndicesForFile]{getIndicesForFile()}},\code{\link[=createMasterScanFile2]{createMasterScanFile2()}}
}
