% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScans.R
\name{createMasterScanFile2}
\alias{createMasterScanFile2}
\title{Create a master scan file which maps scan dependencies in mgf peakists.  For scan sequences with two MS2 files.}
\usage{
createMasterScanFile2(ms2PAVApeaklistCID = NA, ms2PAVApeaklistETD = NA)
}
\arguments{
\item{ms2PAVApeaklistCID}{path to MS2 CID peaklist (parent of MS3 scans).}

\item{ms2PAVApeaklistETD}{path to MS2 ETD peaklist (optional).}
}
\value{
A data frame
}
\description{
\code{createMasterScanFile2()} is for scan sequences that contain two dependent
MS2 scans and no MS3 scans. They are annotted in the help file as MS2-CID and
MS2-ETD peaklists but can be any type. Requires using the PAVA formating conventions
for the peaklists filenames. For instance, PAVA creates separate peaklist
files for each scan type and appends a suffic such as: \verb{_FTMSms2ethcd.txt}
for high-res MS2-EThcD scans, or \verb{_FTMSms2cid.txt} for high-res MS2-CID scans.
The base filename is the same for all the linked peaklist files, corresponding
to the name of the .raw file.
}
\seealso{
\code{\link[=getIndicesForFile]{getIndicesForFile()}},\code{\link[=createMasterScanFile2]{createMasterScanFile2()}}
}
