% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_touchstone}
\alias{dir_touchstone}
\title{The directory where the touchstone database lives}
\usage{
dir_touchstone()
}
\description{
The directory where the touchstone database lives
}
\keyword{internal}
