% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{distancePlot}
\alias{distancePlot}
\title{Histogram of crosslinked distance distributions.
Internal plotting function used by the Shiny interface to Touchstone
Seems to be deprecated}
\usage{
distancePlot(targetDists, randomDists, threshold)
}
\arguments{
\item{targetDists}{vector of c-alpha c-alpha distances}

\item{randomDists}{vector of random c-alpha c-alpha distances}

\item{threshold}{the distance violation threshold.}
}
\value{
No return value
}
\description{
Histogram of crosslinked distance distributions.
Internal plotting function used by the Shiny interface to Touchstone
Seems to be deprecated
}
\keyword{internal}
