% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{distancePlot2}
\alias{distancePlot2}
\title{Histogram of crosslinked distance distributions.
distancePlot2 works with Touchstone datTab and differs as well as the Shiny interface.}
\usage{
distancePlot2(datTab, threshold = 35, binWidth = NA)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{threshold}{the distance violation threshold.}

\item{binWidth}{binwidth for tabulating the histogram}
}
\value{
A ggplot object
}
\description{
Histogram of crosslinked distance distributions.
distancePlot2 works with Touchstone datTab and differs as well as the Shiny interface.
}
