% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{euclideanDistance}
\alias{euclideanDistance}
\title{Measure euclidean distance between Calpha atoms of a crosslinked residue pair}
\usage{
euclideanDistance(XLink.AA.1, PDB.chain.1, XLink.AA.2, PDB.chain.2, parsedPDB)
}
\arguments{
\item{XLink.AA.1}{Sequence index of 1st residue.}

\item{PDB.chain.1}{PDB identifier for 1st residue.}

\item{XLink.AA.2}{Sequence index of 2nd residue.}

\item{PDB.chain.2}{PDB identifier for 2nd residue.}

\item{parsedPDB}{A parsed Touchstone pdb file.}
}
\value{
A numeric.
}
\description{
Measure euclidean distance between Calpha atoms of a crosslinked residue pair
}
\seealso{
\code{\link[=parsePDB]{parsePDB()}}, \code{\link[=euclideanDistanceHomologous]{euclideanDistanceHomologous()}}, \code{\link[=euclideanHelper]{euclideanHelper()}}
}
