% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expr_eval}
\alias{expr_eval}
\title{Evaluate an expression}
\usage{
expr_eval(text)
}
\arguments{
\item{text}{Character vector with code to evaluate.}
}
\description{
Evaluate an expression
}
\keyword{internal}
