% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exprs_eval}
\alias{exprs_eval}
\title{Evaluate an expression}
\usage{
exprs_eval(...)
}
\arguments{
\item{text}{Character vector with code to evaluate.}
}
\value{
The input, parsed and evaluated.
}
\description{
Evaluate an expression
}
\keyword{internal}
