% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{findSeparateThresholds}
\alias{findSeparateThresholds}
\title{Determine separate SVM.score thresholds for inter and intra-protein hits that give the desired FDR}
\usage{
findSeparateThresholds(
  datTab,
  targetER = 0.01,
  minThreshold = -5,
  classifier = "SVM.score",
  errorFUN = calculateFDR.unseparated,
  ...
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{targetER}{Desired error rate (typically FDR).}

\item{minThreshold}{Minimum acceptable score threshold.}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}

\item{errorFUN}{Function used to estimate error.}

\item{...}{Additional params passed by calling function. Typically scalingFactor.}
}
\value{
A list of inter and intra-protein thresholds to give the desired error rate.
}
\description{
\code{findSeparateThresholds()} uses calculated values at each threshold. Differs from
\code{findSeparateThresholdsModelled()} which first models the error functions. Both
of these functions separate inter-protein and intra-protein crosslinks into separate
classes and find two different thresholds that give the desired error rate.
Can be used on any summarization level.
}
\seealso{
\code{\link[=findThreshold]{findThreshold()}}, \code{\link[=findThresholdModelled]{findThresholdModelled()}}, \code{\link[=findSeparateThresholds]{findSeparateThresholds()}}
}
