% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{findSeparateThresholdsModelled}
\alias{findSeparateThresholdsModelled}
\title{Determine separate SVM.score thresholds for inter and intra-protein hits that give the desired FDR using modeled FDR function.}
\usage{
findSeparateThresholdsModelled(
  datTab,
  targetER = 0.01,
  minThreshold = -5,
  scalingFactor = the$decoyScalingFactor,
  plot = T,
  classifier = "SVM.score"
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{targetER}{Desired FDR.}

\item{minThreshold}{Minimum acceptable SVM.score threshold.}

\item{scalingFactor}{k, multiple by which which decoy DB is larger than target DB}

\item{plot}{Show decoy table plot?}

\item{classifier}{Column name in \code{datTab} used as the classifier to rank hits.}
}
\value{
A list of inter- and intra-protein thresholds. Attributes record the
method used and whether the requested FDR was reached for each class.
}
\description{
\code{findSeparateThresholds()} uses calculated values at each threshold. Differs from
\code{findSeparateThresholdsModelled()} which first models the error functions. Both
of these functions separate inter-protein and intra-protein crosslinks into separate
classes and find two different thresholds that give the desired error rate.
Can be used on any summarization level.
}
\seealso{
\code{\link[=findThreshold]{findThreshold()}}, \code{\link[=findThresholdModelled]{findThresholdModelled()}}, \code{\link[=findSeparateThresholds]{findSeparateThresholds()}}
}
