% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{findThreshold}
\alias{findThreshold}
\title{Determine SVM.score threshold to give desired FDR}
\usage{
findThreshold(
  datTab,
  targetER = 0.01,
  minThreshold = -5,
  classifier = "SVM.score",
  errorFUN = calculateFDR.unseparated,
  ...
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{targetER}{Desired error rate (typically FDR).}

\item{minThreshold}{Minimum acceptable threshold.}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}

\item{errorFUN}{Function used to estimate error.}

\item{...}{Additional params passed by calling function. Typically scalingFactor.}
}
\value{
A list with the threshold and calculated FDR at this threshold
}
\description{
\code{findThreshold()} uses calculated values at each threshold. Differs from
\code{findThresholdModelled()} which first models the error functions. Neither
of these functions separate the data into different classes. Can be used on
any summarization level.
}
\seealso{
\code{\link[=findThresholdModelled]{findThresholdModelled()}}, \code{\link[=findSeparateThresholds]{findSeparateThresholds()}}, \code{\link[=findSeparateThresholds]{findSeparateThresholds()}},
}
