% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScans.R
\name{flattenMS3pair}
\alias{flattenMS3pair}
\title{Helper function called by \code{ms3ionFragFind}.  Combined two linked linear MS3
PSMs into a Prospector-Touchstone MS2 crosslinked data frame.}
\usage{
flattenMS3pair(
  ms3CSM1,
  ms3CSM2,
  fraction,
  masterScanNo = NA,
  frag1 = "Alkene",
  frag2 = "Thiol\\\\(Unsaturated\\\\)"
)
}
\arguments{
\item{ms3CSM1}{The first MS3-PSM.}

\item{ms3CSM2}{The second MS3-PSM.}

\item{fraction}{Base name of RAW file.}

\item{masterScanNo}{Scan number of the triggering scan (typically MS2-CID).}

\item{frag1}{Regex to detect the variable modification corresponding to the cleaved crosslinked fragment.}

\item{frag2}{Regex to detect the variable modification corresponding to the other cleaved crosslinked fragment.}
}
\value{
A data frame
}
\description{
Helper function called by \code{ms3ionFragFind}.  Combined two linked linear MS3
PSMs into a Prospector-Touchstone MS2 crosslinked data frame.
}
\seealso{
\code{\link[=ms3ionFragFind]{ms3ionFragFind()}}, \code{\link[=processMS3xlinkResults]{processMS3xlinkResults()}}, \code{\link[=addMasterScanInfo]{addMasterScanInfo()}}
}
