% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{gatherPDBs}
\alias{gatherPDBs}
\title{Downloads and prepares all pdb files specified in the module file.}
\usage{
gatherPDBs(modFile, pdbFileDir = getwd())
}
\arguments{
\item{modFile}{A parsed Touchstone module file.}

\item{pdbFileDir}{Path to local directory containing pdb/cif file}
}
\value{
A list of Calpha coordinates for each pdb id.
}
\description{
Downloads and prepares all pdb files specified in the module file.
}
\seealso{
\code{\link[=parsePDB]{parsePDB()}}
}
