% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{generateErrorTable}
\alias{generateErrorTable}
\title{This is an older version of generateDecoyTable, which is still called a number times.}
\usage{
generateErrorTable(
  datTab,
  classifier = "SVM.score",
  errorFUN = calculateFDR.unseparated,
  ...,
  class.min = NA,
  class.max = NA
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}

\item{errorFUN}{error function used to generate decoy table (eg, FDR or violation rate, or GT)}

\item{...}{additional params passed by calling functions}

\item{class.min}{minimum value of classifier to consider}

\item{class.max}{maximum value of classifier to consider}
}
\value{
A data frame
}
\description{
This is an older version of generateDecoyTable, which is still called a number times.
}
\keyword{internal}
