% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{generateMSViewerLink.ms3}
\alias{generateMSViewerLink.ms3}
\title{Internal function to generate url links to MS-Product spectra. For MS3 data.}
\usage{
generateMSViewerLink.ms3(
  path,
  fraction,
  z,
  peptide,
  spectrum,
  instrumentType = NA,
  outputType = "HTML"
)
}
\arguments{
\item{path}{Base url to MS-Product spectra (typically a path on MS-Viewer for instance)}

\item{fraction}{Fraction identifier}

\item{z}{Precursor charge}

\item{peptide}{Peptide sequence}

\item{spectrum}{spectrum identifier}

\item{instrumentType}{Prospector instrument type parameter}

\item{outputType}{MS-Product output type (typically "text" or "HTML")}
}
\value{
A character containing a uri.
}
\description{
Internal function to generate url links to MS-Product spectra. For MS3 data.
}
\keyword{internal}
