% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{generateMSViewerLink}
\alias{generateMSViewerLink}
\title{Internal function to generate url links to MS-Product spectra. For MS2 data.}
\usage{
generateMSViewerLink(
  path,
  fraction,
  z,
  peptide.1,
  peptide.2,
  spectrum,
  instrumentType = NA,
  linkType = NA,
  outputType = "HTML"
)
}
\arguments{
\item{path}{Base url to MS-Product spectra (typically a path on MS-Viewer for instance)}

\item{fraction}{Fraction identifier}

\item{z}{Precursor charge}

\item{peptide.1}{Peptide 1 sequence}

\item{peptide.2}{Peptide 2 sequence}

\item{spectrum}{spectrum identifier}

\item{instrumentType}{Prospector instrument type parameter}

\item{linkType}{Prospector link type parameter}

\item{outputType}{MS-Product output type (typically "text" or "HTML")}
}
\value{
A character containing a uri.
}
\description{
Internal function to generate url links to MS-Product spectra. For MS2 data.
}
\keyword{internal}
