% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_changed_files}
\alias{get_changed_files}
\title{Get changed files.}
\usage{
get_changed_files(
  refs = c(ref_get_or_fail("GITHUB_BASE_REF"), ref_get_or_fail("GITHUB_HEAD_REF"))
)
}
\arguments{
\item{refs}{Git refs to compare.}
}
\value{
Character Vector of file names or an error object.
}
\description{
description
}
\keyword{internal}
