% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{getDiagnosticMatches}
\alias{getDiagnosticMatches}
\title{Are diagnostic/reporter ions (MH*/#) from cleavable crosslinking reagents detected?}
\usage{
getDiagnosticMatches(msms.ions, msms.int)
}
\arguments{
\item{msms.ions}{Columns containing list of detected ions, must be present in Search Compare output.}

\item{msms.int}{Columns with corresponding intensity values for detected ions.}
}
\value{
A list of diagnostic ions and their intensities.
}
\description{
Are diagnostic/reporter ions (MH*/#) from cleavable crosslinking reagents detected?
}
\seealso{
\code{\link[=calculateDiagnosticPairs]{calculateDiagnosticPairs()}}
}
