% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScans.R
\name{getIndicesForFile}
\alias{getIndicesForFile}
\title{Gathers information from the headers of an MGF formatted peaklist file.}
\usage{
getIndicesForFile(file)
}
\arguments{
\item{file}{Path to the mgf peaklist.}
}
\value{
A data frame
}
\description{
This function assumes PAVA formatted mgf file headers. There is some variability
in the header format depending on the program used to generate the mgf and I
haven't tested other programs output. \code{getIndicesforFile()} is looking for
the Scan No in the \code{TITLE} linke as well as a \code{Master_Scan_Number()}.
}
\seealso{
\code{\link[=createMasterScanFile]{createMasterScanFile()}}, \code{\link[=createMasterScanFile2]{createMasterScanFile2()}}
}
