% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScans.R
\name{getMassError}
\alias{getMassError}
\title{Calculate the mass error between theoretical and measured values in ppm.}
\usage{
getMassError(targetMass, testMass)
}
\arguments{
\item{targetMass}{theoretical mass of peptide.}

\item{testMass}{experimental mass of peptide.}
}
\value{
A numeric value
}
\description{
Calculate the mass error between theoretical and measured values in ppm.
}
