% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{getUniprotSeqLength}
\alias{getUniprotSeqLength}
\title{Returns the length of a protein sequence in number of AA residues.}
\usage{
getUniprotSeqLength(uniprotAcc)
}
\arguments{
\item{uniprotAcc}{A valid Uniprot Accession Number.}
}
\value{
an integer
}
\description{
Returns the length of a protein sequence in number of AA residues.
}
