% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{hash_pkg}
\alias{hash_pkg}
\title{When did the package sources change last?}
\usage{
hash_pkg(path_pkg)
}
\arguments{
\item{path_pkg}{The path to the repository to install.}
}
\description{
When did the package sources change last?
}
\keyword{internal}
