% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{install_missing_deps}
\alias{install_missing_deps}
\title{Install missing BASE dependencies}
\usage{
install_missing_deps(path_pkg, quiet = FALSE)
}
\description{
If the HEAD branch removes dependencies (compared to BASE), installing the
package from the BASE branch will fail due to missing dependencies, as
dependencies were installed in the GitHub Action based on the \code{DESCRIPTION}
of the HEAD branch. The simplest way to
ensure all required dependencies are present (including specification of
remotes) is by simply installing them
into the respective {touchstone} library. Prepend the local touchstone
library to the library path with \code{\link[=local_touchstone_libpath]{local_touchstone_libpath()}}.
}
\keyword{internal}
