% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_installed}
\alias{is_installed}
\title{Check if a package is installed and unloading it}
\usage{
is_installed(path_pkg = ".")
}
\description{
Check if a package is installed and unloading it
}
\keyword{internal}
