% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{lengthFilter}
\alias{lengthFilter}
\title{Prefilter crosslinks or CSMs by length of each peptide.}
\usage{
lengthFilter(datTab, minLen, maxLen)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{minLen}{Mimum peptide length, as number of amino acid residues.}

\item{maxLen}{Maximum peptide length, as number of amino acid residues.}
}
\value{
A data frame
}
\description{
Prefilter crosslinks or CSMs by length of each peptide.
}
