% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{local_clean_touchstone}
\alias{local_clean_touchstone}
\title{Clean up}
\usage{
local_clean_touchstone(envir = parent.frame())
}
\arguments{
\item{envir}{\verb{[environment]}\cr Attach exit handlers to this environment.
Typically, this should be either the current environment or
a parent frame (accessed through \code{\link[=parent.frame]{parent.frame()}}).}
}
\description{
Deletes \code{\link[=dir_touchstone]{dir_touchstone()}} when the local frame is destroyed.
}
\seealso{
Other testers: 
\code{\link{local_package}()}
}
\concept{testers}
\keyword{internal}
