% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{local_package}
\alias{local_package}
\title{Create a test package}
\usage{
local_package(
  path = path_temp_pkg("testpkg"),
  branches = c("main", "devel"),
  envir = parent.frame()
)
}
\arguments{
\item{path}{The path to the temporary package.}

\item{branches}{Branches to be created.}

\item{envir}{\verb{[environment]}\cr Attach exit handlers to this environment.
Typically, this should be either the current environment or
a parent frame (accessed through \code{\link[=parent.frame]{parent.frame()}}).}
}
\description{
Creates a package in a temporary directory until the local frame is
destroyed.
}
\seealso{
Other testers: 
\code{\link{local_clean_touchstone}()}
}
\concept{testers}
