% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_tempdir_setwd}
\alias{local_tempdir_setwd}
\title{Temporarily set the working directory to a temp directory}
\usage{
local_tempdir_setwd(.local_envir = parent.frame())
}
\description{
Temporarily set the working directory to a temp directory
}
\keyword{internal}
