% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{local_test_setup}
\alias{local_test_setup}
\title{Combines a few options for a good testing environment}
\usage{
local_test_setup(envir = parent.frame())
}
\arguments{
\item{envir}{The environment that determines when the deferred actions will
manifest.}

\item{keywords}{internal}
}
\description{
Combines a few options for a good testing environment
}
\details{
Setup includes:
\itemize{
\item temporarily disables usethis communication.
\item runs touchstone for only two iterations for speeding things up.
\item remove the touchstone env with \code{\link[=local_clean_touchstone]{local_clean_touchstone()}}.
}
}
\keyword{internal}
