% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{local_touchstone_libpath}
\alias{local_touchstone_libpath}
\title{Set Library Path}
\usage{
local_touchstone_libpath(branch, env = parent.frame())
}
\arguments{
\item{branch}{Git branch to use, e.g. HEAD or BASE branch.}

\item{env}{Environment in which the change should be applied.}
}
\description{
Temporarily add a touchstone library to the path, so it can be found by
\code{\link[=.libPaths]{.libPaths()}} and friends. Can be used in \link{touchstone_script}
to prepare benchmarks etc. If there are touchstone libraries on the path
when this function is called, they will be removed.
}
\seealso{
\code{\link[=run_script]{run_script()}}
}
\keyword{internal}
