% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_without_touchstone_lib}
\alias{local_without_touchstone_lib}
\title{Temporarily remove all touchstone libraries from the path}
\usage{
local_without_touchstone_lib(path_pkg = ".", envir = parent.frame())
}
\arguments{
\item{path_pkg}{The path to the package that contains the touchstone library.}

\item{envir}{The environment that triggers the deferred action on
destruction.}
}
\description{
This is useful in conjunction with \code{\link[=run_script]{run_script()}}.
}
\details{
\itemize{
\item Add a touchstone library to the path with \code{\link[=run_script]{run_script()}} and
run a script. The script hence may contain calls to libraries only installed
in touchstone libraries.
\item benchmark code with \code{\link[=benchmark_run]{benchmark_run()}}. At the start, remove all
all touchstone libraries from path and add the touchstone library we need.
}

Advantages: Keep benchmarked repo in touchstone library only.
}
\keyword{internal}
