% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{massErrorPlot}
\alias{massErrorPlot}
\title{Histogram of precursor mass errors.
Internal plotting function used by the Shiny interface to Touchstone}
\usage{
massErrorPlot(massErrors, lowThresh, highThresh, lowPlotRange, highPlotRange)
}
\arguments{
\item{massErrors}{precursor mass accuracy error (in ppm)}

\item{lowThresh}{mass accuracy threshold, low (in ppm)}

\item{highThresh}{mass accuracy threshold, high (in ppm)}

\item{lowPlotRange}{x-axis limit, low (in ppm)}

\item{highPlotRange}{x-axis limit, high (in ppm)}
}
\value{
No return value
}
\description{
Histogram of precursor mass errors.
Internal plotting function used by the Shiny interface to Touchstone
}
\keyword{internal}
