% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{measureCrosslinkDistance}
\alias{measureCrosslinkDistance}
\title{Measure crosslinked and randomized distributions of euclidean distances for a dataset.}
\usage{
measureCrosslinkDistance(datTabMod, pdbList)
}
\arguments{
\item{datTabMod}{Parsed CLMS search results, with modules and chain ids assigned.}

\item{pdbList}{A list of pdb ids.}
}
\value{
a data frame
}
\description{
For each pdb id assigned in the module file, euclidean crosslinked and randomized
Calpha-Calpha distance distributions are measured and assigned to the dataset.
}
\seealso{
\code{\link[=processModuleFile]{processModuleFile()}}, \code{\link[=euclideanHelper]{euclideanHelper()}}, \code{\link[=randomHelper]{randomHelper()}}
}
