% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{mmin}
\alias{mmin}
\title{Internal convenience function for determining parameter ranges and bin sizes.}
\usage{
mmin(x, base = 5)
}
\arguments{
\item{x}{numeric parameter}

\item{base}{desired subdivision (integer)}
}
\value{
numeric
}
\description{
Internal convenience function for determining parameter ranges and bin sizes.
}
\keyword{internal}
