% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{moduleTilePlot}
\alias{moduleTilePlot}
\title{Plots the CSMs tabulated by Module-Pair in a Tile Plot}
\usage{
moduleTilePlot(
  datTab,
  threshold = -100,
  title = "Module Plot",
  modBorders = NULL,
  ...
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{threshold}{Score threshold for classifying data.}

\item{title}{Plot title.}

\item{modBorders}{Numeric vector to specify optional borders between modules.}

\item{...}{Parameters passed to \code{summarizeModuleData()}}
}
\value{
A ggplt object
}
\description{
Plots the CSMs tabulated by Module-Pair in a Tile Plot
}
\seealso{
\code{\link[=summarizeModuleData]{summarizeModuleData()}}
}
