% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScans.R
\name{ms3ionFragFind}
\alias{ms3ionFragFind}
\title{Takes linked MS3-level PSMs and tests whether there are complementary scans
that sum up to the precursor mass to detect putative crosslinks. This is
developmental and currently only works with DSSO. Requires more testing.
Requires the MS3-search results to have \verb{Master Scan File} information appended.}
\usage{
ms3ionFragFind(fraction, masterScanNo, scanGroupTibble, tol = 25, mis = 0)
}
\arguments{
\item{fraction}{Base name of RAW file.}

\item{masterScanNo}{Scan number of the triggering scan (typically MS2-CID).}

\item{scanGroupTibble}{Nested ms3 level linear search results.}

\item{tol}{precursor Mass tolerance in ppm.}

\item{mis}{misannotated precursor ion index.}
}
\value{
A data frame
}
\description{
Takes linked MS3-level PSMs and tests whether there are complementary scans
that sum up to the precursor mass to detect putative crosslinks. This is
developmental and currently only works with DSSO. Requires more testing.
Requires the MS3-search results to have \verb{Master Scan File} information appended.
}
\seealso{
\code{\link[=processMS3xlinkResults]{processMS3xlinkResults()}}, \code{\link[=addMasterScanInfo]{addMasterScanInfo()}}
}
