% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{numHitsPlot}
\alias{numHitsPlot}
\title{Plot of number of crosslinked hits by FDR.
Internal plotting function used by the Shiny interface to Touchstone}
\usage{
numHitsPlot(num.hits, threshold = 0)
}
\arguments{
\item{num.hits}{A data frame created by \code{generateErrorTable()}}

\item{threshold}{FDR threshold to plot}
}
\value{
A ggplot object
}
\description{
Plot of number of crosslinked hits by FDR.
Internal plotting function used by the Shiny interface to Touchstone
}
\seealso{
\code{\link[=generateErrorTable]{generateErrorTable()}}
}
