% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{pairPlot}
\alias{pairPlot}
\title{Contact map style dot plot showing crosslinked residue pairs}
\usage{
pairPlot(
  datTab,
  modTab = NULL,
  color = "lightseagreen",
  accOrder = NULL,
  removeMods = NA_character_,
  displayEmpty = T
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{modTab}{A parsed Touchstone module file.}

\item{color}{Named or hex coded color for dots.}

\item{accOrder}{Character vector of Accession numbers in their desired plot order.}

\item{removeMods}{Modules to remove from plot.}

\item{displayEmpty}{Draw module boxes even when they contain no crosslinks.#'}
}
\value{
the ggplot object
}
\description{
Function takes a datTab which is reduced to Residue Pairs and plot
Dot for each unique crosslink with size equal to num of instances. The
color can reflect the source of the crosslink to facilitate comparisons.
Basic skeleton of the function that will get built up later. Ultimately
would like this to work on the XLSearchOutput object, which probably needs
flag to show if it is residue pair reduced etc.
}
