% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{parseCrosslinker}
\alias{parseCrosslinker}
\title{Skeleton function to parse heavy / light crosslink pairs.}
\usage{
parseCrosslinker(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
Currently only works for D12/H12 BS3
}
