% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{parsePDB}
\alias{parsePDB}
\title{Prepare PDB file for crosslinked distance measurements}
\usage{
parsePDB(pdbID = NA, pdbFileDir = getwd())
}
\arguments{
\item{pdbID}{Code for pdb/cif file}

\item{pdbFileDir}{Path to local directory containin pdb/cif file}
}
\value{
A data frame.
}
\description{
Searches for pdb file in the pdbFileDir and if it fails to find it there,
downloads from the pdb repository. Works with either pdb or mmCIF formatted
files. Retains only Calpha coordinates. Wrapper for functions in the \code{Bio3d} package.
}
