% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{path_pinned_asset}
\alias{path_pinned_asset}
\title{Get path to asset}
\usage{
path_pinned_asset(..., ref = ref_get_or_fail("GITHUB_HEAD_REF"))
}
\arguments{
\item{...}{character vectors, if any values are NA, the result will also be
NA. The paths follow the recycling rules used in the tibble package,
namely that only length 1 arguments are recycled.}

\item{ref}{Currently without function as assets can only be oinned from the
head branch. Introduced for API stability.}
}
\value{
The absolute path to the asset.
}
\description{
This function is used to get the absolut path to a pinned asset
within \code{script.R}.
}
