% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pin_head_asssets}
\alias{pin_head_asssets}
\title{Pin asset directory}
\usage{
pin_head_asssets(...)
}
\arguments{
\item{...}{A number of directories, as strings in relation to the current
working directory, that contain scripts you want to source in \code{script.R}.}
}
\value{
The temp dir invisibly.
}
\description{
Add directories that need to be available on both branches
when running\code{script.R}. During \link{benchmark_run_ref} they will be placed
in the same directory as \code{script.R}.
}
\examples{
\dontrun{
# In script.R
pin_head_asssets(c("bench", "inst/scripts"))

source("scripts/setup.R")

touchstone::benchmark_run_ref(
  expr_before_benchmark = {
    !!setup
    source("bench/exprs.R")
  },
  run_me = some_exprs(),
  n = 6
)
}
}
