% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{pr_comment}
\alias{pr_comment}
\title{Modifying the PR Comment}
\description{
The files \code{touchstone/header.R} and \code{touchstone/footer.R} allow you to modify
the PR comment. The files will be evaluated in the context of
\code{\link[=benchmarks_analyze]{benchmarks_analyze()}} and should return one string containing the text.
You can use github markdown e.g. emojis like :tada: in the string.
}
\section{Header}{

Available variables for glue substitution:
\itemize{
\item ci: confidence interval
\item branches: BASE and HEAD branches benchmarked against each other.
}
}

\section{Footer}{

There are no special variables available in the footer.
You can access the benchmark results via \code{\link[=path_pr_comment]{path_pr_comment()}}.
}

\seealso{
\code{\link[base:eval]{base::eval()}} \code{\link[base:parse]{base::parse()}}
}
