% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppiContextCalibration.R
\name{print.touchstone_ppi_results}
\alias{print.touchstone_ppi_results}
\title{Print contextual protein-pair results}
\usage{
\method{print}{touchstone_ppi_results}(x, ...)
}
\arguments{
\item{x}{A \code{touchstone_ppi_results} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print contextual protein-pair results
}
