% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print.touchstone_results}
\alias{print.touchstone_results}
\title{Print prepared crosslink results}
\usage{
\method{print}{touchstone_results}(x, ...)
}
\arguments{
\item{x}{A result returned by \code{\link[=prepareCrosslinkResults]{prepareCrosslinkResults()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print prepared crosslink results
}
