% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{processModuleFile}
\alias{processModuleFile}
\title{Wrapper function that assigns modules and distances defined in a Touchstone Module File}
\usage{
processModuleFile(datTab, pathToModFile, pdbFileDir = getwd())
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{pathToModFile}{Location of module file}

\item{pdbFileDir}{Path to local directory containing pdb/cif file}
}
\value{
a data frame
}
\description{
Wrapper function that assigns modules and distances defined in a Touchstone Module File
}
