% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScans.R
\name{processMS3xlinkResults}
\alias{processMS3xlinkResults}
\title{Recombines linked MS3-search results into crosslinked Peptides.}
\usage{
processMS3xlinkResults(ms3results, tol = 25)
}
\arguments{
\item{ms3results}{Parsed MS3-search results.}

\item{tol}{precursor Mass tolerance in ppm.}
}
\value{
A data frame.
}
\description{
\code{processMS3xlinkResults()} takes Protein Prospector search results for linear
peptide from MS3-scans of cleavable crosslinkers and searches links scans for
possible crosslinked spectral matches (CSMs). This is developmental and currently
only works with DSSO. Requires more testing. Requires the MS3-search results
to have \verb{Master Scan File} information appended.
}
\seealso{
\code{\link[=ms3ionFragFind]{ms3ionFragFind()}}, \code{\link[=addMasterScanInfo]{addMasterScanInfo()}}
}
