% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{productIonFilter}
\alias{productIonFilter}
\title{exclude CSMs below a given number of bond cleavages matched per peptide.}
\usage{
productIonFilter(datTab, minProducts.1 = 3, minProducts.2 = 3)
}
\arguments{
\item{minProducts.1}{number of bonds cleaved in peptide 1}

\item{minProducts.2}{number of bonds cleaved in peptide 2}
}
\description{
exclude CSMs below a given number of bond cleavages matched per peptide.
}
