% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{randomCrosslinks}
\alias{randomCrosslinks}
\title{Sample random C-alpha distances on a PDB structure.}
\usage{
randomCrosslinks(
  parsedPDB,
  numCrosslinks,
  resType1 = "LYS",
  resType2 = "LYS",
  chains = NULL
)
}
\arguments{
\item{parsedPDB}{A parsed Touchstone pdb file.}

\item{numCrosslinks}{Number of random crosslinks generated.}

\item{resType1}{A character vector containing three letter amino acid codes, from which the random crosslinks are selected.}

\item{resType2}{A character vector containing three letter amino acid codes, from which the random crosslinks are selected.}

\item{chains}{If not \code{null}, restrict the random distribution to certain chains, given in a character vector.}
}
\value{
A numeric vector.
}
\description{
Generates a distribution of random C-alpha, C-alpha distances on a PDB structure.
The size of the distibution is given my \code{numCrosslinks}. By default, only Lys-Lys
pairs are sampled, but this can be changed with the \code{resType} parameters.
}
\seealso{
\code{\link[=randomHelper]{randomHelper()}}
}
