% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{randomHelper}
\alias{randomHelper}
\title{Generate random distributions of crosslinked euclidean distances across a dataset.}
\usage{
randomHelper(
  datTab,
  parsedPDB,
  resType1 = "LYS",
  resType2 = "LYS",
  chains = NULL
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{parsedPDB}{A parsed Touchstone pdb file.}

\item{resType1}{A character vector containing three letter amino acid codes, from which the random crosslinks are selected.}

\item{resType2}{A character vector containing three letter amino acid codes, from which the random crosslinks are selected.}

\item{chains}{If not \code{null}, restrict the random distribution to certain chains, given in a character vector.}
}
\value{
A data frame.
}
\description{
Generate random distributions of crosslinked euclidean distances across a dataset.
}
\seealso{
\code{\link[=randomCrosslinks]{randomCrosslinks()}}
}
