% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{readModuleFile}
\alias{readModuleFile}
\title{Parse a Touchstone formatted module file.}
\usage{
readModuleFile(pathToModFile)
}
\arguments{
\item{pathToModFile}{Location of module file.}
}
\value{
A data frame.
}
\description{
Parse a Touchstone formatted module file.
}
