% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedScans.R
\name{readMS3results}
\alias{readMS3results}
\title{Reads Prospector formatted search compare output from MS3 searches of linear peptides.}
\usage{
readMS3results(ms3SearchTable)
}
\arguments{
\item{ms3SearchTable}{Search Compare Output from a linear Protein Prospector linear peptide search.}
}
\value{
A data frame.
}
\description{
This is a developmental function for generating results from MS3-based search
strategies using cleavable linkers.
}
