% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{readParamsFile}
\alias{readParamsFile}
\title{Reads the Protein Prospector parameter xml files.}
\usage{
readParamsFile(paramsFile)
}
\arguments{
\item{paramsFile}{xml file with the parameters submitted to Batch-Tag}
}
\value{
An xml_document
}
\description{
Reads the Protein Prospector parameter xml files.
}
\keyword{internal}
