% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{readProspectorXLOutput}
\alias{readProspectorXLOutput}
\title{Read Search Compare Output from Protein Prospector}
\usage{
readProspectorXLOutput(
  inputFile,
  minPepLen = 3,
  minPepScore = 0,
  minScoreDiff = 0,
  minIons = 3
)
}
\arguments{
\item{inputFile}{A tab delimited search compare file generated using tstoneMS2.json param file}

\item{minPepLen}{minimum length of each peptide in amino acids.}

\item{minPepScore}{minimum prospector score of each peptide.}

\item{minScoreDiff}{minimum Score Difference for the CSM.}

\item{minIons}{mimimum number of product ions matched for each peptide.}
}
\description{
Read Search Compare Output from Protein Prospector
}
