% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{ref_install}
\alias{ref_install}
\title{Checks out a source branch and install the package}
\usage{
ref_install(ref = "master", path_pkg = ".", install_dependencies = FALSE)
}
\arguments{
\item{ref}{A reference to a git commit. Currently, only branch names are
supported.}

\item{path_pkg}{The path to the repository to install.}

\item{install_dependencies}{Passed to \code{\link[remotes:install_local]{remotes::install_local()}}. Set to
\code{FALSE} can help when ran locally without internet connection.}
}
\value{
A character vector with library paths.
}
\description{
The installed source package is cached. An installation is forced on source
code change or version number.
}
\keyword{internal}
