% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{removeDecoys}
\alias{removeDecoys}
\title{Remove decoy hits from tables}
\usage{
removeDecoys(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
Convenience function to remove decoys, typically before producing a report.
}
