% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{removeModule}
\alias{removeModule}
\title{Removes all crosslinks from a given module from the dataset.}
\usage{
removeModule(datTab, modules)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{modules}{character vector of modules to remove.}
}
\value{
A data frame
}
\description{
Removes all crosslinks from a given module from the dataset.
}
