% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{renameProtein}
\alias{renameProtein}
\title{Replace accession number in search results.}
\usage{
renameProtein(datTab, oldName, newName)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{oldName}{Accession number to replace.}

\item{newName}{New accession number.}
}
\value{
a data frame
}
\description{
Replace accession number in search results.
}
