% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{renumberProtein}
\alias{renumberProtein}
\title{Renumber protein in search results.}
\usage{
renumberProtein(datTab, protein, shift)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{protein}{Accession number to renumber.}

\item{shift}{Difference between old sequence position and new sequence position}
}
\value{
a data frame
}
\description{
This would typically be used when the sequence of a protein construct used in
an experiment differs from the UniProt reference, due to an epitope tag or deletion,
and one would like to harmonize the results reporting.
}
