% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_on_change}
\alias{run_on_change}
\title{Run benchmark on file change}
\usage{
run_on_change(
  files,
  benchmark,
  refs = c(ref_get_or_fail("GITHUB_BASE_REF"), ref_get_or_fail("GITHUB_HEAD_REF"))
)
}
\arguments{
\item{files}{Character vector with paths to files to check for change.}

\item{benchmark}{Call to \link{benchmark_run_ref}.}

\item{refs}{Git refs to compare.}
}
\description{
The benchmark will only run if \emph{any} file of \code{files} was modified
compared to \code{ref}. This allows for shorter commit messages.
}
\examples{
\dontrun{
touchstone::run_on_change(
  "R/core.R",
  touchstone::benchmark_run_ref(
    expr_before_benchmark = source(c("dir/data.R", "dir/another.R")),
    random_test = yourpkg::f(),
    n = 2
  )
)
}
}
