% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{scoreFilter}
\alias{scoreFilter}
\title{Prefilter crosslinks or CSMs by Prospector Score of each peptide.}
\usage{
scoreFilter(datTab, minScore = 0)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{minScore}{Minimum Protein Prospector Score for each individual peptide}
}
\value{
A data frame
}
\description{
Prefilter crosslinks or CSMs by Prospector Score of each peptide.
}
