% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{set_lib_path}
\alias{set_lib_path}
\title{Set Library Path}
\usage{
set_lib_path(ref, env = parent.frame())
}
\arguments{
\item{ref}{Git ref to use, e.g. HEAD or BASE ref.}

\item{env}{Environment in which the change should be applied.}
}
\description{
Prefixes \code{\link[=.libPaths]{.libPaths()}} with the library containing the \code{ref}
version of the package. Can be used in \link{touchstone_script}
to prepare benchmarks etc..
}
\seealso{
\code{\link[=run_script]{run_script()}}
}
