% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{setDecoyScalingFactor}
\alias{setDecoyScalingFactor}
\title{Set the decoy scaling factor}
\usage{
setDecoyScalingFactor(dsf)
}
\arguments{
\item{dsf}{an integer, the decoy scaling factor k.}
}
\value{
No reutrn value.
}
\description{
The decoy scaling factor is an integer k that describes how how much larger the
decoy database is to the target database.  Should be set globally once for each
analysis, after loading or reloading Touchstone. All functions with a
\code{scalingFactor} argument use this value by default. Defaults to 1 whenever the
package is loaded.
}
