% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{summarizeModuleData}
\alias{summarizeModuleData}
\title{Tabulates crosslink spectral counts (CSMs) by Module Pair, typically for plotting.
Used by interactive, Shiny Touchstone interface.}
\usage{
summarizeModuleData(datTab, clearDiag = F, modOrder = NULL)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{clearDiag}{Whether to de-duplicate Module Pairs, for plotting only half of symmetrical matrix.}

\item{modOrder}{Optional character vector showing order that modules should be arranged/plotted.}
}
\value{
A data frame
}
\description{
Tabulates crosslink spectral counts (CSMs) by Module Pair, typically for plotting.
Used by interactive, Shiny Touchstone interface.
}
