% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{summarizeProtData}
\alias{summarizeProtData}
\title{Tabulates crosslink spectral counts (CSMs) by Protein Pair, typically for plotting.
Used by interactive, Shiny Touchstone interface.}
\usage{
summarizeProtData(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
Tabulates crosslink spectral counts (CSMs) by Protein Pair, typically for plotting.
Used by interactive, Shiny Touchstone interface.
}
