% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{touchstone_example}
\alias{touchstone_example}
\title{Convenience function for working through the ribosome example dataset.}
\usage{
touchstone_example(path = NULL)
}
\arguments{
\item{path}{path to system file}
}
\value{
A character vector
}
\description{
Convenience function for working through the ribosome example dataset.
}
