% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{touchstone_managers}
\alias{touchstone_managers}
\alias{dir_touchstone}
\alias{touchstone_clear}
\title{Touchstone managers}
\usage{
dir_touchstone()

touchstone_clear()
}
\description{
Utilities to manage the touchstone database.
}
\section{Functions}{
\itemize{
\item \code{dir_touchstone}: returns the directory where the touchstone database lives.

\item \code{touchstone_clear}: clears the touchstone database.
}}

