% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{trainClassifier}
\alias{trainClassifier}
\title{Automated function to select features, build SVM score, and perform hyper-parameter tuning}
\usage{
trainClassifier(
  datTab,
  params = NULL,
  scoreName = "SVM.score",
  scalingFactor = the$decoyScalingFactor,
  targetER = 0.01,
  preFilterER.values = c(0.45, 0.35, 0.25)
)
}
\arguments{
\item{datTab}{Parsed CLMS search results}

\item{params}{Character vector specifying names of the features in \code{datTab} used to train model.}

\item{scoreName}{Name for the new scoring function.}

\item{scalingFactor}{An integer k. The multiple by which decoy DB is larger than target DB}

\item{targetER}{Desired FDR for classification of CSMs}

\item{preFilterER.values}{Numeric vector of preFilter FDR values for hyperparamter tuning}
}
\value{
A list object containing the preFiltered data with the SVM.score as well as other information about the training procedure.
}
\description{
replaced by buildCrosslinkScore but keeping this function around for backward capability.
\code{trainClassifier()} is the main function that most users will use to automatically
used for model training and then calls \code{buildClassifier()} at different pre-filter
values and selects the best performing model.
}
\seealso{
\code{\link[=trainClassifier_parallel]{trainClassifier_parallel()}}, \link{buildClassifier}
}
