% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{trainClassifier_parallel}
\alias{trainClassifier_parallel}
\title{Automated function to select features, build SVM score, and perform hyper-paramter tuning}
\usage{
trainClassifier_parallel(
  datTab,
  params = NA,
  scoreName = "SVM.score",
  scalingFactor = the$decoyScalingFactor,
  targetER = 0.01,
  preFilterER.values = c(0.45, 0.35, 0.25)
)
}
\arguments{
\item{datTab}{Parsed CLMS search results}

\item{params}{Character vector specifying names of the features in \code{datTab} used to train model.}

\item{scoreName}{Name for the new scoring function.}

\item{scalingFactor}{An integer k. The multiple by which decoy DB is larger than target DB}

\item{targetER}{Desired FDR for classification of CSMs}

\item{preFilterER.values}{Numeric vector of preFilter FDR values for hyperparamter tuning}
}
\value{
A list object containing the preFiltered data with the SVM.score as well as other information about the training procedure.
}
\description{
\code{trainClassifier_parallel()} is the parallelized version \code{trainClassifier()}.
Requires that \code{furrr} package to be installed and user must select an appropriate
\code{furrr::plan()}. In testing with large datasets, the parallelized computation
does not results in significant improvements in processing time. This function is
still under development.
}
\seealso{
\code{\link[=trainClassifier_parallel]{trainClassifier_parallel()}}, \link{buildClassifier}
}
