% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_touchstone_workflows}
\alias{use_touchstone_workflows}
\title{Use touchstone GitHub Actions Workflows}
\usage{
use_touchstone_workflows(
  overwrite = FALSE,
  on_comment = FALSE,
  command = "/benchmark",
  limit_to = c("OWNER", "MEMBER", "COLLABORATOR"),
  force_upstream = TRUE
)
}
\arguments{
\item{overwrite}{Overwrites files if they exist.}

\item{on_comment}{Don't benchmark every commit. Only trigger on PR
comments starting with \code{command}.}

\item{command}{The command to use to trigger the benchmark with a comment.}

\item{limit_to}{Roles that are allowed to trigger the benchmark workflow
via comment. See details for a list of roles and their definition.
Set to \code{NULL} to allow everyone to trigger a benchmark.}

\item{force_upstream}{Always benchmark against the upstream base branch.}
}
\value{

}
\description{
This function will add (or update) the {touchstone} GitHub Actions workflows
to your package repository. Use in the root directory of your repository.
This function will be called by \code{\link[=use_touchstone]{use_touchstone()}}, you should
only need to call it to update the workflows or change their parameters.
}
\details{
Possible roles for \code{limit_to}:
\itemize{
\item \code{OWNER}: Owner of the repository, e.g. user for user/repo.
\itemize{
\item It is undocumented who holds this status in an org.
}
\item \code{MEMBER}: Member of org for org/repo.
\item \code{COLLABORATOR}: Anyone who was added as a collaborator to a repository.
\item \code{CONTRIBUTOR}: Anyone who has contributed any commit to the repository.
}

Each user has only one role and the check does not interpolate permissions,
so you have to add all roles whom you want to have permission to start the
benchmark. So if you only add "COLLABORATOR" the owner will not be able to
start the benchmark.

GitHub will recognize additional, mostly unusual roles, see the
\href{https://docs.github.com/en/rest/issues/comments}{documentation}.
}
