% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{violationRate}
\alias{violationRate}
\title{Calculate the violation rate}
\usage{
violationRate(datTab, threshold, classifier = "Score.Diff", maxDistance = 35)
}
\arguments{
\item{datTab}{parsed CLMS search results}

\item{threshold}{single score threshold}

\item{classifier}{which classifier to use to rank hits}

\item{maxDistance}{Calpha-Calpha distance which is considered a violation (Å)}
}
\value{
A numeric value
}
\description{
Calculate the violation rate
}
