% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{with_touchstone_lib}
\alias{with_touchstone_lib}
\title{Sources a script}
\usage{
with_touchstone_lib(path, ref = Sys.getenv("GITHUB_HEAD_REF"))
}
\arguments{
\item{path}{The script to run. It must fulfill the requirements of a
\link{touchstone_script}.}

\item{ref}{The ref that corresponds to the library that should be prepended
to the library path when the script at \code{path} is executed. Note that during
a benchmark run with \code{\link[=benchmark_run_ref]{benchmark_run_ref()}}, all touchstone library paths
are removed and the library path corresponding to the argument \code{ref}
in \code{\link[=benchmark_run_ref]{benchmark_run_ref()}} is added.}
}
\description{
Basically \code{\link[base:source]{base::source()}}, but prepending the library path with a
touchstone library and running the script in a temp directory to avoid
git operations like checking out different branches to interfere with the
script execution (as running the script changes itself through git checkout).
}
