% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{XVisOutput}
\alias{XVisOutput}
\title{Puts crosslink results into XVis format for visualization.
see Grimm et al, NAR 2017.
https://doi.org/10.1093/nar/gkv463}
\usage{
XVisOutput(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
Puts crosslink results into XVis format for visualization.
see Grimm et al, NAR 2017.
https://doi.org/10.1093/nar/gkv463
}
