% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{activate}
\alias{activate}
\alias{deactivate}
\title{Activate touchstone environment}
\usage{
activate(
  head_ref = gert::git_branch(),
  base_ref = getOption("touchstone.default_base_ref", "main"),
  env = parent.frame()
)

deactivate(env = parent.frame())
}
\arguments{
\item{head_ref}{Git ref to be used as the HEAD ref when running benchmarks.
Defaults to the current branch.}

\item{base_ref}{Git ref for the baselin ref when running benchmarks. Defaults
'main' if the option \code{touchstone.default_base_ref}is not set.}

\item{env}{In which environment the temporary changes should be made.
For use within functions.}
}
\description{
This sets envvars and options to make interactivley running
\link{touchstone_script} possible.
}
\section{Functions}{
\itemize{
\item \code{deactivate}: Restore the original environment state.
}}

\examples{
\dontrun{
activate()
# You can now test parts of your touchstone/script.R
deactivate()
}
}
