% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{activate}
\alias{activate}
\alias{deactivate}
\title{Activate touchstone environment}
\usage{
activate(
  head_ref = gert::git_branch(),
  base_ref = getOption("touchstone.default_base_ref", "main"),
  env = parent.frame()
)

deactivate(env = parent.frame())
}
\arguments{
\item{head_ref}{Git ref to be used as the \code{GITHUB_HEAD_REF} ref (i.e. the
branch with new changes) when running benchmarks. Defaults to the current
branch.}

\item{base_ref}{Git ref for the \code{GITHUB_BASE_REF} (i.e. the branch you want
to merge your changes into) when running benchmarks.
Defaults to 'main' if the option \code{touchstone.default_base_ref}is not set.}

\item{env}{In which environment the temporary changes should be made.
For use within functions.}
}
\description{
This sets environment variables, R options and library paths to work
interactively on the \link{touchstone_script}.
}
\section{Functions}{
\itemize{
\item \code{deactivate}: Restore the original environment state.
}}

\examples{
\dontrun{
activate()
# You can now test parts of your touchstone script, e.g. touchstone/script.R
deactivate()
}
}
