% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{addXlinkCount}
\alias{addXlinkCount}
\title{Count number of CSMs per URP and the number of URPs per PP}
\usage{
addXlinkCount(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results}
}
\value{
A data frame
}
\description{
Count number of CSMs per URP and the number of URPs per PP
}
