% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_no_global_installation}
\alias{assert_no_global_installation}
\title{Make sure there is no installation of the package to benchmark in the global
package library}
\usage{
assert_no_global_installation(path_pkg = ".")
}
\description{
Make sure there is no installation of the package to benchmark in the global
package library
}
\keyword{internal}
