% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{assignXLinkClass}
\alias{assignXLinkClass}
\title{is the crosslink interprotein (hetromeric) or intraprotein (homomeric)}
\usage{
assignXLinkClass(datTab)
}
\arguments{
\item{datTab}{}
}
\description{
is the crosslink interprotein (hetromeric) or intraprotein (homomeric)
}
