% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{benchmark_ls}
\alias{benchmark_ls}
\title{List which benchmarks were recorded}
\usage{
benchmark_ls(name = "")
}
\arguments{
\item{name}{The name of the benchmark.}
}
\value{
A tibble with name and refs of the existing benchmarks.
}
\description{
List which benchmarks were recorded
}
