% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{benchmark_read}
\alias{benchmark_read}
\title{Read benchmarks}
\usage{
benchmark_read(name, ref)
}
\arguments{
\item{name}{The name of the benchmark.}

\item{ref}{A character vector of length one to indicate the git ref (i.e.
commit, tag, branch etc) of the benchmarking.}
}
\value{
A tibble with the benchmarks.
}
\description{
Read benchmarks
}
