% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{benchmark_run_iteration}
\alias{benchmark_run_iteration}
\title{Run a benchmark iteration}
\usage{
benchmark_run_iteration(
  expr_before_benchmark,
  ...,
  ref,
  n = getOption("touchstone.n_iterations", 20)
)
}
\arguments{
\item{expr_before_benchmark}{Character vector with code to run before
the benchmark is ran, will be evaluated with \code{\link[=exprs_eval]{exprs_eval()}}.}

\item{...}{Named character vector of length one with code to benchmark, will
be evaluated with \code{\link[=exprs_eval]{exprs_eval()}}.}

\item{ref}{A character vector of length one to indicate the git ref (i.e.
commit, tag, branch etc) of the benchmarking.}

\item{n}{The number of times to run a benchmark, each time with a separate
call to \code{\link[bench:mark]{bench::mark()}}.}
}
\description{
Run a benchmark iteration
}
