% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{benchmark_run_ref_impl}
\alias{benchmark_run_ref_impl}
\title{Checkout a branch from a repo and run an iteration}
\usage{
benchmark_run_ref_impl(ref, expr_before_benchmark, ..., path_pkg)
}
\description{
Checkout a branch from a repo and run an iteration
}
\keyword{internal}
