% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{benchmark_write}
\alias{benchmark_write}
\title{Write a benchmark}
\usage{
benchmark_write(
  benchmark,
  name,
  ref,
  block = NA,
  iteration = NA,
  append = TRUE
)
}
\arguments{
\item{benchmark}{The result of \code{\link[bench:mark]{bench::mark()}}, with \code{iterations = 1}.}

\item{name}{The name of the benchmark.}

\item{ref}{A character vector of length one to indicate the git ref (i.e.
commit, tag, branch etc) of the benchmarking.}

\item{block}{All refs appear once in a block.}

\item{iteration}{An integer indicating to which iteration the benchmark
refers to. Multiple iterations within a block always benchmark the same
\code{ref}.}

\item{append}{Whether to append the result to the file or not.}
}
\value{
Character vector of lenth one with path to the record written (invisibly).
}
\description{
Write a benchmark
}
