% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{benchmarks_analyze}
\alias{benchmarks_analyze}
\title{Turn raw benchmark results into text and figures}
\usage{
benchmarks_analyze(
  refs = c(Sys.getenv("GITHUB_BASE_REF"), Sys.getenv("GITHUB_HEAD_REF"))
)
}
\arguments{
\item{refs}{The names of the branches for which analysis should be created.}
}
\description{
Turn raw benchmark results into text and figures
}
\details{
Creates two side effects:
\itemize{
\item Density plots for each element in \code{refs} are written to \code{touchstone/plots}.
\item A text explaining the speed diff is written to
\code{touchstone/pr-comment/info.txt} for every registered benchmarking
expression.
}
}
