% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{bestResPair}
\alias{bestResPair}
\title{Summarize on Unique Residue Pairs (URPs)}
\usage{
bestResPair(datTab, classifier = .data$SVM.score, retainGroups = T)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}

\item{retainGroups}{Will retain pre-existing group in the data frame when apply the summarization.}
}
\value{
A data frame
}
\description{
The default behaviour is to summarize residue-pairs within existing groups.
This would typically be different experimental conditions. To overwrite and
calculate the residue-pairs on the entire data table seq the \code{retainGroup} option
to \code{False}.
}
\seealso{
\code{\link[=bestPepPair]{bestPepPair()}}, \code{\link[=bestProtPair]{bestProtPair()}}, \code{\link[=bestModPair]{bestModPair()}}
}
