% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{buildSVM}
\alias{buildSVM}
\title{Basic function to build a new SVM classifier.  Doesn't do any feature selection or
hyperparamter tuning. Build two separate SVM models on different subsets of the data
and averages the results.}
\usage{
buildSVM(
  datTab,
  params = params.best,
  scoreName = "SVM.score",
  sampleNo = 20000,
  showTab = F,
  ...
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{params}{Character vector specifying names of the features in \code{datTab} used to train model.}

\item{scoreName}{Name for the new scoring function.}

\item{sampleNo}{Size of the training dataset (integer).}

\item{showTab}{print classificaiton table?}

\item{...}{paramters passed to \code{e1071:svm()} function}
}
\value{
A data frame, one column larger than the input containing the new score.
}
\description{
Basic function to build a new SVM classifier.  Doesn't do any feature selection or
hyperparamter tuning. Build two separate SVM models on different subsets of the data
and averages the results.
}
\seealso{
\code{\link[=trainCrosslinkScore]{trainCrosslinkScore()}}, \code{\link[=tuneSVM.helper]{tuneSVM.helper()}}, \code{\link[=tuneSVM]{tuneSVM()}}
}
