% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{cache_up_to_date}
\alias{cache_up_to_date}
\alias{cache_update}
\alias{cache_get}
\title{Cache package sources within a session}
\usage{
cache_up_to_date(branch, path_pkg)

cache_update(branch, path_pkg)

cache_get()
}
\description{
This is required to make sure \code{\link[remotes:install_local]{remotes::install_local()}} installs again
when source code changed.
}
\keyword{internal}
