% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculateDecoys}
\alias{calculateDecoys}
\title{determine whether a CSM is a decoy or target}
\usage{
calculateDecoys(datTab)
}
\arguments{
\item{datTab}{parsed CLMS search results}
}
\description{
determine whether a CSM is a decoy or target
}
