% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculateDiagnosticPairs}
\alias{calculateDiagnosticPairs}
\title{determined whether diagnostic ion pair from cleavable crosslinker was deteceted}
\usage{
calculateDiagnosticPairs(datTab, int.thresh = 0.05)
}
\arguments{
\item{int.thresh}{minimum intensity for diagnostic ion.}
}
\description{
determined whether diagnostic ion pair from cleavable crosslinker was deteceted
}
