% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculateDiagnosticPairsNonCleavable}
\alias{calculateDiagnosticPairsNonCleavable}
\title{analog of calcualte diagnostic pairs for non-cleavable crosslinkers}
\usage{
calculateDiagnosticPairsNonCleavable(datTab)
}
\arguments{
\item{datTab}{}
}
\description{
analog of calcualte diagnostic pairs for non-cleavable crosslinkers
}
