% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculatePairs}
\alias{calculatePairs}
\title{Determine residue, peptide,and protein pairs}
\usage{
calculatePairs(datTab, scalingFactor = the$decoyScalingFactor)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{scalingFactor}{An integer describing how many times larger the decoy
database is than the target database.}
}
\value{
A data frame
}
\description{
Determine residue, peptide,and protein pairs
}
