% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculatePairs}
\alias{calculatePairs}
\title{determine residue, peptide,and protein pairs}
\usage{
calculatePairs(datTab)
}
\arguments{
\item{datTab}{}
}
\description{
determine residue, peptide,and protein pairs
}
