% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculatePairs}
\alias{calculatePairs}
\title{Determine residue, peptide,and protein pairs}
\usage{
calculatePairs(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
Determine residue, peptide,and protein pairs
}
