% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculatePeptideLengths}
\alias{calculatePeptideLengths}
\title{calculate length of each peptide in amino acid residues.}
\usage{
calculatePeptideLengths(datTab)
}
\arguments{
\item{datTab}{}
}
\description{
calculate length of each peptide in amino acid residues.
}
