% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{calculateProductIons}
\alias{calculateProductIons}
\title{calculates number of distinct bonds cleavages observed as product ions.}
\usage{
calculateProductIons(datTab)
}
\arguments{
\item{datTab}{}
}
\description{
calculates number of distinct bonds cleavages observed as product ions.
}
