% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{classifyDataset}
\alias{classifyDataset}
\title{Classify CLMS datasets}
\usage{
classifyDataset(datTab, threshold = list(), classifier = .data$SVM.score)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{threshold}{A list or a numeric of length 1. List must contain either \code{globalThresh} or both \code{interThresh} and \code{intraThresh} numeric elements.}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}
}
\value{
A data frame
}
\description{
The \code{threshold} (a list) argument determines the behavior of \code{classifyDataset}.
If threshold contains a \code{globalThresh} element than \code{classifySingleThreshold()}
is called. If \code{threshold} contains \code{interThresh} and \code{intraThresh} elemenst than
inter-protein and intra-protein hits are treated separately and \code{classifySeparateThresholds()}
is called. \code{threshold} can also be a numeric value which call \code{classifySingleThreshold()}
}
\seealso{
\code{\link[=classifySeparateThresholds]{classifySeparateThresholds()}}, \code{\link[=classifySingleThreshold]{classifySingleThreshold()}}
}
