% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{classifySeparateThresholds}
\alias{classifySeparateThresholds}
\title{Classify CLMS datasets using separate interProtein and intraProtein thresholds}
\usage{
classifySeparateThresholds(
  datTab,
  separateThresh,
  classifier = .data$SVM.score
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{separateThresh}{interProtein and intraProtein score thresholds as list(intraThresh = x, interThresh = y)}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}
}
\value{
A data frame
}
\description{
Classify CLMS datasets using separate interProtein and intraProtein thresholds
}
\seealso{
\code{\link[=classifyDataset]{classifyDataset()}}, \code{\link[=classifySingleThreshold]{classifySingleThreshold()}}
}
