% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{classifySingleThreshold}
\alias{classifySingleThreshold}
\title{Classify CLMS dataset using a single threshold}
\usage{
classifySingleThreshold(datTab, singleThresh, classifier = .data$SVM.score)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{singleThresh}{Either a list that contains \code{globalThresh} numeric elements or a single numeric value.}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}
}
\value{
A data frame
}
\description{
Classify CLMS dataset using a single threshold
}
\seealso{
\code{\link[=classifySeparateThresholds]{classifySeparateThresholds()}}, \code{\link[=classifyDataset]{classifyDataset()}}
}
