% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{countDecoys}
\alias{countDecoys}
\title{Convenience function to display the confusion matrix.}
\usage{
countDecoys(datTab, threshold = NULL)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{threshold}{A list or a numeric of length 1. List must contain either \code{globalThresh} or both \code{interThresh} and \code{intraThresh} numeric elements.}
}
\value{
A data frame
}
\description{
Convenience function to display the confusion matrix.
}
