% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{deScaler}
\alias{deScaler}
\title{Rescales data to plot and tabulate estimate decoy hits}
\usage{
deScaler(datTab, scalingFactor = the$decoyScalingFactor)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{scalingFactor}{The decoy scaling factor}
}
\value{
A data frame
}
\description{
Called by the FDRplots function, but also useful when tabulating results.
For cases where a larger decoy database is searched to help estimate incorrect hits.
The decoy hits are divided by the scaling factor so that we tabulate the estimated
error rather than the artifically large number of decoys.
}
