% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{euclideanDistanceHomologous}
\alias{euclideanDistanceHomologous}
\title{Measure euclidean distance between Calpha atoms of a crosslinked residue pair when there are homologous results.}
\usage{
euclideanDistanceHomologous(
  XLink.AA.1,
  PDB.chain.1,
  XLink.AA.2,
  PDB.chain.2,
  parsedPDB
)
}
\arguments{
\item{XLink.AA.1}{Sequence index of 1st residue.}

\item{PDB.chain.1}{PDB identifier for 1st residue.}

\item{XLink.AA.2}{Sequence index of 2nd residue.}

\item{PDB.chain.2}{PDB identifier for 2nd residue.}

\item{parsedPDB}{A parsed Touchstone pdb file.}
}
\value{
A numeric.
}
\description{
\code{euclideanDistanceHomologous()} expects that all possible chain ids are supplied as
a comma separated character string. For instance is the first crosslinked residue can
be either position 100 on either chain A, D, G, \code{XLink.AA.1} value should be \code{100} and
the \code{PDB.chain.1} should be given as \code{"A, D, G"}. All possible euclidean distances
are evaluated, and the shortest distance is reported. This function is meant to
handle cases in which a protein complex has multiple copies of the same subunit,
rather than situations where the crosslinked peptide appears in multiple different
homologous proteins.
}
\seealso{
\code{\link[=parsePDB]{parsePDB()}}, \code{\link[=euclideanDistance]{euclideanDistance()}}, \code{\link[=euclideanHelper]{euclideanHelper()}}
}
