% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distModFile.R
\name{euclideanHelper}
\alias{euclideanHelper}
\title{Measure crosslinked euclidean distances over the entire results file.}
\usage{
euclideanHelper(datTab, parsedPDB)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{parsedPDB}{A parsed Touchstone pdb file.}
}
\value{
A data frame.
}
\description{
Measure crosslinked euclidean distances over the entire results file.
}
\seealso{
\code{\link[=parsePDB]{parsePDB()}}, \code{\link[=euclideanDistanceHomologous]{euclideanDistanceHomologous()}}
}
