% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exprs_eval}
\alias{exprs_eval}
\title{Evaluate an expression for sideeffects}
\usage{
exprs_eval(..., env = rlang::caller_env())
}
\arguments{
\item{...}{Character vector  of length 1 or expression with code to evaluate. This will be quoted using
\code{\link[rlang:nse-defuse]{rlang::enexprs()}}, so if you want to use \code{quote}ed expressions you have to unquote them using \verb{!!}}

\item{env}{Environment in which the expression will be evaluated.}
}
\value{
The input (invisibly).
}
\description{
Evaluate an expression for sideeffects
}
\keyword{internal}
