% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{fdrPlots}
\alias{fdrPlots}
\title{Plots the score distributions of decoy and target crosslinked hits.}
\usage{
fdrPlots(
  datTab,
  threshold = 0,
  classifier = .data$SVM.score,
  scalingFactor = the$decoyScalingFactor,
  separateFacets = T,
  addLegend = T,
  title = "FDR plot"
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{threshold}{Score threshold for classifying data.}

\item{classifier}{Classifier to use.}

\item{scalingFactor}{The decoy scaling factor.}

\item{separateFacets}{Whether to plot intraProtein and interProtein hits in separate facets}

\item{addLegend}{Whether to display the legend.}

\item{title}{Title to plot.}
}
\value{
A ggplot object.
}
\description{
Plots the score distributions of decoy and target crosslinked hits.
}
