% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{findThresholdModelled}
\alias{findThresholdModelled}
\title{Determine SVM.score threshold to give desired FDR, using modeled FDR function}
\usage{
findThresholdModelled(
  datTab,
  targetER = 0.01,
  minThreshold = -5,
  scalingFactor = the$decoyScalingFactor,
  ...,
  classifier = "SVM.score"
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{targetER}{Desired FDR.}

\item{minThreshold}{Minimum acceptable SVM.score threshold.}

\item{scalingFactor}{An integer k. The multiple by which decoy DB is larger than target DB.}

\item{...}{Additional params passed by calling function.}

\item{classifier}{Column name in \code{datTab} used as the classifier to rank hits.}
}
\value{
A list with the threshold and calculated FDR at this threshold
}
\description{
\code{findThreshold()} uses calculated values at each threshold. Differs from
\code{findThresholdModelled()} which first models the error functions. Neither
of these functions separate the data into different classes. Can be used on
any summarization level.
}
\seealso{
\code{\link[=findThreshold]{findThreshold()}}, \code{\link[=findSeparateThresholds]{findSeparateThresholds()}}
}
