% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{formatXLTable}
\alias{formatXLTable}
\title{Formats the crosslink results for exporting / viewing}
\usage{
formatXLTable(datTab, msviewer = F, extraCols = NULL)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{msviewer}{Formats column names to be compatible with MS-Viewer}

\item{extraCols}{Character vector of extra column names to be included in the report}
}
\value{
A data frame
}
\description{
Formats the crosslink results for exporting / viewing
}
