% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{generateDecoyTable}
\alias{generateDecoyTable}
\title{Model dependence of FDR on scoring function}
\usage{
generateDecoyTable(
  datTab,
  targetER = 0.01,
  scalingFactor = the$decoyScalingFactor,
  classifier = "SVM.score",
  plot = T
)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{targetER}{desired error rate (typically FDR, but can be something else like violation rate or ground truth based error)}

\item{scalingFactor}{An integer k. The multiple by which decoy DB is larger than target DB.}

\item{classifier}{Column name in \code{datTab} used as the classifier to use to rank hits.}

\item{plot}{Show plot?}
}
\value{
A data frame
}
\description{
Internal function called by other functions in \code{fdr.R}
}
\keyword{internal}
