% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{getDiagnosticMatches}
\alias{getDiagnosticMatches}
\title{are diagnostic MH*/# ions from cleavable crosslinking reagents detected?}
\usage{
getDiagnosticMatches(msms.ions, msms.int)
}
\arguments{
\item{msms.ions}{columns containing list of detected ions, must be present in Search Compare output.}

\item{msms.int}{columns with corresponding intensity vaues fro detected ions.}
}
\description{
are diagnostic MH*/# ions from cleavable crosslinking reagents detected?
}
