% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{getDiagnosticMatchesNonCleavable}
\alias{getDiagnosticMatchesNonCleavable}
\title{Analog of getDiagnostic Matches for non-cleavable crosslinkers. Looks for P, PL, PLK type ions.}
\usage{
getDiagnosticMatchesNonCleavable(msms.ions, msms.int)
}
\arguments{
\item{msms.int}{}
}
\description{
Analog of getDiagnostic Matches for non-cleavable crosslinkers. Looks for P, PL, PLK type ions.
}
