% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{getProductIonMatches}
\alias{getProductIonMatches}
\title{parses the list matched product ions and translate to peptide bond position.}
\usage{
getProductIonMatches(msms.ions, pep.len)
}
\arguments{
\item{msms.ions}{list of product ion matches found in Search Compare output}

\item{pep.len}{length of peptide}
}
\description{
parses the list matched product ions and translate to peptide bond position.
}
