% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{getStringScores}
\alias{getStringScores}
\title{Download STRING-DB scores for interProtein Crosslinks.}
\usage{
getStringScores(datTab, ncbiTaxonomyCode = NULL)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{ncbiTaxonomyCode}{NCBI format species code used by string-db}
}
\value{
A data frame
}
\description{
Call the STRINGdb library and queries string for evidence supporting the
crosslinked PPIs. Attempts to automatically detect the species code from the
CLMS results, but this will only work for the most common research organisms.
Otherwise the user should provide the ncbiTaxonomy code:
\verb{https://string-db.org/} under the \code{Oraganisms} link.
}
