% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_installed}
\alias{is_installed}
\title{Check if a package is installed and unloading it}
\usage{
is_installed(path_pkg = ".")
}
\value{
A list with \code{name} of the package to check and \code{installed}, a boolean to
indicate if the package is installed or not.
}
\description{
Check if a package is installed and unloading it
}
\keyword{internal}
