% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{lengthFilter}
\alias{lengthFilter}
\title{prefilter CSMs by length of each peptide.}
\usage{
lengthFilter(datTab, minLen, maxLen)
}
\arguments{
\item{maxLen}{}
}
\description{
prefilter CSMs by length of each peptide.
}
