% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{local_package}
\alias{local_package}
\title{Create a test package}
\usage{
local_package(
  pkg_name = fs::path_file(fs::file_temp("pkg")),
  branches = c("main", "devel"),
  r_sample = NULL,
  setwd = TRUE,
  envir = parent.frame()
)
}
\arguments{
\item{pkg_name}{The name of the temporary package.}

\item{branches}{Branches to be created.}

\item{r_sample}{Character with code to write to \code{R/sampleR.}. This is helpful
to validate if the installed package corresponds to source branch for
testing. If \code{NULL}, nothing is written.}

\item{setwd}{Whether or not the working directory should be temporarily
set to the package root.}

\item{envir}{\verb{[environment]}\cr Attach exit handlers to this environment.
Typically, this should be either the current environment or
a parent frame (accessed through \code{\link[=parent.frame]{parent.frame()}}).}
}
\description{
Creates a package in a temporary directory and sets the working directory
until the local frame is destroyed.
}
\details{
This is primarily for testing.
}
\seealso{
Other testers: 
\code{\link{local_clean_touchstone}()}
}
\concept{testers}
\keyword{internal}
