% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingAndhelper.R
\name{makeXiNetFile}
\alias{makeXiNetFile}
\title{Puts crosslink results into XiNet format for visualization.
see Combe et al, MCP 2015.}
\usage{
makeXiNetFile(datTab)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}
}
\value{
A data frame
}
\description{
Puts crosslink results into XiNet format for visualization.
see Combe et al, MCP 2015.
}
