% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pin_assets}
\alias{pin_assets}
\title{Pin asset directory}
\usage{
pin_assets(..., ref = ref_get_or_fail("GITHUB_HEAD_REF"), overwrite = TRUE)
}
\arguments{
\item{...}{Any number of directories or files, as strings, that you want to
access in \code{script.R}.}

\item{ref}{The branch the passed assets are copied from.}
}
\value{
The asset directory invisibly.
}
\description{
Pin files or directories that need to be available on both branches when
running\code{script.R}. During \link{benchmark_run_ref} they will available via
\link{path_pinned_asset}.
}
\details{
When passing nested directories or files within nested directories
only the file/last directory will be copied. Directories will be copied
recursively. See examples.
}
\examples{
\dontrun{
# In script.R

pin_assets(c("bench", "inst/setup.R", "some/nested/dir"))

source(path_pinned_asset("setup.R"))
load(path_pinned_asset("dir/data.RData"))

touchstone::benchmark_run_ref(
  expr_before_benchmark = {
    !!setup
    source(path_pinned_asset("bench/exprs.R"))
  },
  run_me = some_exprs(),
  n = 6
)
}
}
