% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{productIonFilter}
\alias{productIonFilter}
\title{Exclude crosslinks or CSMs below a threshold number of bond cleavages matched per peptide.}
\usage{
productIonFilter(datTab, minProducts.1 = 3, minProducts.2 = 3)
}
\arguments{
\item{datTab}{Parsed CLMS search results.}

\item{minProducts.1}{Number of bonds cleaved in peptide 1.}

\item{minProducts.2}{Number of bonds cleaved in peptide 2.}
}
\value{
A data frame
}
\description{
Exclude crosslinks or CSMs below a threshold number of bond cleavages matched per peptide.
}
\seealso{
\code{\link[=calculateProductIons]{calculateProductIons()}}, \code{\link[=getProductIonMatches]{getProductIonMatches()}}
}
