% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{readProspectorXLOutput}
\alias{readProspectorXLOutput}
\title{Read Search Compare Output from Protein Prospector}
\usage{
readProspectorXLOutput(
  inputFile,
  minPepLen = 3,
  minPepScore = 0,
  minScoreDiff = 0,
  minIons = 0
)
}
\arguments{
\item{inputFile}{A tab delimited search compare file generated using tstoneMS2.json param file}

\item{minPepLen}{Minimum length of each peptide in amino acids.}

\item{minPepScore}{Minimum Prospector score of each peptide.}

\item{minScoreDiff}{Minimum Score Difference for the CSM.}

\item{minIons}{Mimimum number of product ions matched for each peptide.}
}
\value{
A data frame
}
\description{
Function for importing Crosslinked Search Compare output into Touchstone.
Search Compare output should be saved in tab delimited text format ideally
created with tstone specific param files, such as \code{tstoneMS2.json}.
}
