% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{refs_install}
\alias{refs_install}
\title{Install branches}
\usage{
refs_install(
  refs = c(Sys.getenv("GITHUB_BASE_REF", abort_missing_ref()),
    Sys.getenv("GITHUB_HEAD_REF", abort_missing_ref())),
  path_pkg = ".",
  install_dependencies = FALSE
)
}
\arguments{
\item{refs}{The names of the branches in a character vector.}

\item{path_pkg}{The path to the repository to install.}

\item{install_dependencies}{Passed to ``}
}
\value{
The global and touchstone library paths
}
\description{
Installs each \code{ref} in a separate library for isolation.
}
