% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processInput.R
\name{scoreFilter}
\alias{scoreFilter}
\title{prefilter CSMs by Prospector Score of each peptide}
\usage{
scoreFilter(datTab, minScore = 0)
}
\arguments{
\item{minScore}{}
}
\description{
prefilter CSMs by Prospector Score of each peptide
}
