% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touchstone-package.R
\docType{package}
\name{touchstone-package}
\alias{touchstone}
\alias{touchstone-package}
\title{touchstone: Benchmarking Different 'Git' Refs with Confidence in One Go}
\description{
A common problem of benchmarking in continuous integration is
    that the computational power of the virtual machines that run the job
    varies over time. This package runs benchmarking code of two 'Git'
    refs of the same repository in a random sequence for better
    comparison, either locally or via 'GitHub Actions' and outputs a
    confidence interval for the relative difference.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lorenzwalthert/touchstone/}
  \item \url{https://lorenzwalthert.github.io/touchstone/}
  \item Report bugs at \url{https://github.com/lorenzwalthert/touchstone/issues/}
}

}
\author{
\strong{Maintainer}: Lorenz Walthert \email{lorenz.walthert@icloud.com}

Authors:
\itemize{
  \item Jacob Wujciak-Jens \email{jacob@wujciak.de} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID})
}

}
\keyword{internal}
