% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touchstone-package.R
\docType{package}
\name{touchstone-package}
\alias{touchstone}
\alias{touchstone-package}
\title{touchstone: Benchmark different git refs in one run}
\description{
A common problem of benchmarking in continuous
    integration is that the computational power of the virtual machines
    that run the job varies over time.  This package allows users to
    benchmark two git refs of the same repo in a random sequence for
    better comparison.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
