% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/touchstone-package.R
\docType{package}
\name{touchstone-package}
\alias{touchstone}
\alias{touchstone-package}
\title{touchstone: Benchmark Different Git Refs in One Run}
\description{
A common problem of benchmarking in continuous integration is
    that the computational power of the virtual machines that run the job
    varies over time.  This package allows users to benchmark two git refs
    of the same repo in a random sequence for better comparison.
}
\author{
\strong{Maintainer}: Lorenz Walthert \email{lorenz.walthert@icloud.com}

}
\keyword{internal}
