% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{touchstone_managers}
\alias{touchstone_managers}
\alias{dir_touchstone}
\alias{touchstone_clear}
\alias{path_pr_comment}
\title{Touchstone managers}
\usage{
dir_touchstone()

touchstone_clear(all = FALSE)

path_pr_comment()
}
\arguments{
\item{all}{Whether to clear the whole touchstone directory or just the
records sub directory.}
}
\value{
Character vector of length one with the path to the touchstone directory (for
\code{dir_touchstone()}), path to the deleted files for (\code{touchstone_clear()}).
}
\description{
Utilities to manage the touchstone database.
}
\section{Functions}{
\itemize{
\item \code{dir_touchstone}: returns the directory where the touchstone database lives.

\item \code{touchstone_clear}: clears the touchstone database.

\item \code{path_pr_comment}: returns the path to the file containing the pr comment.
}}

\seealso{
\link{pr_comment}
}
