% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{touchstone_script}
\alias{touchstone_script}
\title{The script for benchmarking}
\description{
The script that contains the code which executes the benchmark. It is
typically called with \code{\link[=run_script]{run_script()}}.
}
\section{Requirements}{


A touchstone script must:
\itemize{
\item install all versions of the benchmarked repository with \code{\link[=branches_install]{branches_install()}}.
\item create benchmarks with one or more calls to \code{\link[=benchmark_run]{benchmark_run()}}.
\item produce the artifacts required in the GitHub workflow with
\code{\link[=benchmarks_analyze]{benchmarks_analyze()}}.
}
}

