% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{trainCrosslinkScore}
\alias{trainCrosslinkScore}
\title{Trains an SVM classifier for CLMS datasets. Performs basic feature selection with
different sets of features used for small database searches and larger one.
Performs some prefilitering of the data that optimizes performance based on database
search size. Performs hyperparamater optimization in a grid search using \code{tuneSVM()}
and chooses the model which produces the most inter-protein crosslinked residue pairs
in the range of 0.01 and 0.05 FDR.}
\usage{
trainCrosslinkScore(
  datTab,
  params = NULL,
  scoreName = "SVM.score",
  scalingFactor = the$decoyScalingFactor,
  targetER = 0.01,
  sampleNo = 20000,
  cost_values = c(1, 5, 10),
  gamma_values = c(0.01, 0.05, 0.1),
  sd_values = c(0, 5, 10, 15, 20)
)
}
\arguments{
\item{datTab}{Parsed CLMS search results}

\item{params}{Character vector specifying names of the features in \code{datTab} used to train model.}

\item{scoreName}{Name for the new scoring function.}

\item{scalingFactor}{An integer k. The multiple by which decoy DB is larger than target DB}

\item{targetER}{Desired FDR for classification of CSMs}

\item{sampleNo}{Size of the training dataset (integer).}

\item{cost_values}{Numeric vector of cost values used for hyperparameter tuning of the SVM model}

\item{gamma_values}{Numeric vector of gamma values used for hyperparameter tuning of the SVM model}

\item{sd_values}{Numeric vector of Score Diff values to use for prefilitering optimiziation.}
}
\value{
A list containing all of the SVM models at different cost and gamma values
as well as a summary table.
}
\description{
Trains an SVM classifier for CLMS datasets. Performs basic feature selection with
different sets of features used for small database searches and larger one.
Performs some prefilitering of the data that optimizes performance based on database
search size. Performs hyperparamater optimization in a grid search using \code{tuneSVM()}
and chooses the model which produces the most inter-protein crosslinked residue pairs
in the range of 0.01 and 0.05 FDR.
}
\seealso{
\code{\link[=tuneSVM]{tuneSVM()}}, \code{\link[=tuneSVM.helper]{tuneSVM.helper()}}, \code{\link[=buildSVM]{buildSVM()}}
}
