% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{tuneSVM.helper}
\alias{tuneSVM.helper}
\title{Helper function called by \code{tuneSVM()} that in turn calls \code{buildSVM()} and
calculates intermediate values in determining the objective functions used
for optimization. Namely, the errorTable generated by \code{generateErrorTable.sep()}
that reports number of interProtein and intraProtein crosslinks by FDR threshold.}
\usage{
tuneSVM.helper(
  datTab,
  targetER = 0.01,
  params = params.best,
  scoreName = "SVM.score",
  scalingFactor = the$decoyScalingFactor,
  sampleNo = 20000,
  cost,
  gamma,
  kernel,
  seed = 1,
  splitBy = NULL
)
}
\arguments{
\item{datTab}{Parsed CLMS search results}

\item{targetER}{Desired FDR for classification of URPss}

\item{params}{Character vector specifying names of the features in \code{datTab} used to train model.}

\item{scoreName}{Name for the new scoring function.}

\item{scalingFactor}{An integer k. The multiple by which decoy DB is larger than target DB}

\item{sampleNo}{Size of the training dataset (integer).}

\item{cost}{Cost value passed to \code{e1071::svm()}}

\item{gamma}{Gamma value passed to \code{e1071::svm()}}

\item{kernel}{Kernel value passed to \code{e1071::svm()}}

\item{seed}{Integer seed used to make cross-fitting reproducible.}

\item{splitBy}{Character vector naming columns whose rows must remain together
during cross-fitting.}
}
\value{
A list containing the trained data at CSM and URP levels, score thresholds
for the targetER, the error table and some other information used for tuning.
}
\description{
Helper function called by \code{tuneSVM()} that in turn calls \code{buildSVM()} and
calculates intermediate values in determining the objective functions used
for optimization. Namely, the errorTable generated by \code{generateErrorTable.sep()}
that reports number of interProtein and intraProtein crosslinks by FDR threshold.
}
\seealso{
\code{\link[=trainCrosslinkScore]{trainCrosslinkScore()}}, \code{\link[=tuneSVM]{tuneSVM()}}, \code{\link[=buildSVM]{buildSVM()}}
}
