% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{tuneSVM}
\alias{tuneSVM}
\title{Performs hyperparamter optimziation for SVM model building by calling \code{tuneSVM.helper()}
across the specified grid of cost and gamma values.}
\usage{
tuneSVM(
  datTab,
  params = params.best,
  scoreName = "SVM.score",
  scalingFactor = the$decoyScalingFactor,
  targetER = 0.01,
  sampleNo = 20000,
  cost_values = c(0.5, 1, 5, 10),
  gamma_values = c(0.01, 0.05, 0.1, 0.5)
)
}
\arguments{
\item{datTab}{Parsed CLMS search results}

\item{params}{Character vector specifying names of the features in \code{datTab} used to train model.}

\item{scoreName}{Name for the new scoring function.}

\item{scalingFactor}{An integer k. The multiple by which decoy DB is larger than target DB}

\item{targetER}{Desired FDR for classification of CSMs}

\item{sampleNo}{Size of the training dataset (integer).}

\item{cost_values}{Numeric vector of cost values used for hyperparameter tuning of the SVM model}

\item{gamma_values}{Numeric vector of gamma values used for hyperparameter tuning of the SVM model}
}
\value{
A list containing all of the SVM models at different cost and gamma values.
}
\description{
Performs hyperparamter optimziation for SVM model building by calling \code{tuneSVM.helper()}
across the specified grid of cost and gamma values.
}
\seealso{
\code{\link[=trainCrosslinkScore]{trainCrosslinkScore()}}, \code{\link[=tuneSVM.helper]{tuneSVM.helper()}}, \code{\link[=buildSVM]{buildSVM()}}
}
