% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_touchstone}
\alias{use_touchstone}
\title{Initiate {touchstone}}
\usage{
use_touchstone(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Overwrites files if they exist.}
}
\value{
The function is called for its side effects and returns \code{NULL} (invisibly).
}
\description{
This function will initialize {touchstone} in your package repository, use
from root directory.
}
\details{
For more information see the 'Using touchstone' vignette:
`vignette("touchstone", package = "touchstone")
}
