% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{use_touchstone}
\alias{use_touchstone}
\title{Initiate {touchstone}}
\usage{
use_touchstone(cancel = TRUE)
}
\arguments{
\item{cancel}{Whether or not to also introduce a GitHub Actions
\href{https://github.com/marketplace/actions/cancel-workflow-action}{cancel workflow}
for canceling old runs when new ones are started. This makes sense because
touchstone runs can take a lot of time and compute resources and you usually
don't care about old runs when you pushed new code.}
}
\value{
The function is called for its side effects and returns \code{NULL} (invisibly).
}
\description{
Initiate {touchstone}
}
