% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{with_touchstone_lib}
\alias{with_touchstone_lib}
\title{Sources a script}
\usage{
with_touchstone_lib(path, ref = Sys.getenv("GITHUB_HEAD_REF"))
}
\arguments{
\item{path}{The script to run.}

\item{ref}{The ref that corresponds to the library that should be prepended
to the library path.}
}
\description{
Basically \code{\link[base:source]{base::source()}}, but prepending the library path with a
touchstone library. For isolation, it's not allowed to install the
benchmarked package into the global library, as asserted with
\code{\link[=assert_no_global_installation]{assert_no_global_installation()}}. However, this implies that the package
is not available in the touchstone script outside of benchmark runs. To
remove this limitation, we prepend a touchstone library location that
contains the installed benchmarked package, and temporarily remove it during
benchmarking if another touchstone library is used.
}
